@@ -1,69 +1,75 @@
 {
-  "name": "caprover",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "start": "node ./bin/www",
-    "dev": "echo 'RECOMPILING' && npx madge --circular --extensions ts ./ && rm -rf ./built && npx tsc && sudo ./dev-scripts/dev-reset-service.sh",
-    "clean": "sudo ./dev-scripts/dev-clean-run-as-dev.sh",
-    "tslint": "tslint -c tslint.json -p tsconfig.json",
-    "build": "rm -rf ./built && npx tsc",
-    "test": "rm -rf ./built && npx tsc && jest"
-  },
-  "dependencies": {
-    "@types/bcryptjs": "^2.4.2",
-    "@types/configstore": "^4.0.0",
-    "@types/cookie-parser": "^1.4.1",
-    "@types/dockerode": "^2.5.20",
-    "@types/ejs": "^2.6.3",
-    "@types/express": "^4.17.0",
-    "@types/fs-extra": "^8.0.0",
-    "@types/http-proxy": "^1.17.0",
-    "@types/jest": "^24.0.17",
-    "@types/js-base64": "^2.3.1",
-    "@types/jsonwebtoken": "^8.3.3",
-    "@types/morgan": "^1.7.36",
-    "@types/multer": "^1.3.8",
-    "@types/on-finished": "^2.3.1",
-    "@types/request": "^2.48.2",
-    "@types/require-from-string": "^1.2.0",
-    "@types/serve-favicon": "^2.2.30",
-    "@types/ssh2": "^0.5.38",
-    "@types/tar": "^4.0.3",
-    "@types/uuid": "^3.4.5",
-    "@types/validator": "^10.11.2",
-    "bcryptjs": "^2.4.3",
-    "body-parser": "^1.19.0",
-    "configstore": "^5.0.0",
-    "cookie-parser": "~1.4.4",
-    "debug": "~4.1.1",
-    "dockerode": "^2.5.8",
-    "ejs": "^2.6.2",
-    "express": "^4.17.1",
-    "fs-extra": "^8.1.0",
-    "http-proxy": "^1.17.0",
-    "is-valid-path": "^0.1.1",
-    "js-base64": "^2.5.1",
-    "jsonwebtoken": "^8.5.1",
-    "moment": "^2.24.0",
-    "morgan": "^1.9.1",
-    "multer": "^1.4.2",
-    "on-finished": "^2.3.0",
-    "public-ip": "^3.2.0",
-    "recursive-readdir": "^2.2.2",
-    "request": "^2.88.0",
-    "require-from-string": "^2.0.2",
-    "serve-favicon": "~2.5.0",
-    "simple-git": "^1.124.0",
-    "ssh2": "^0.8.5",
-    "tar": "^4.4.10",
-    "tslint": "^5.18.0",
-    "typescript": "^3.2.4",
-    "uuid": "^3.3.2",
-    "validator": "^11.1.0"
-  },
-  "devDependencies": {
-    "jest": "^24.8.0",
-    "ts-jest": "^24.0.2"
-  }
+    "name": "caprover",
+    "version": "0.0.0",
+    "private": true,
+    "scripts": {
+        "start": "node ./built/server.js",
+        "dev": "npm run build && sudo ./dev-scripts/dev-reset-service.sh",
+        "clean": "npm run build && sudo ./dev-scripts/dev-clean-run-as-dev.sh",
+        "tslint": "tslint -c tslint.json -p tsconfig.json",
+        "formatter": "prettier --check './src/**/*.ts'",
+        "formatter-write": "prettier --write './src/**/*.ts'",
+        "build": "echo 'RECOMPILING' && npx madge --circular --extensions ts ./ && rm -rf ./built && npx tsc && echo 'Build successful'",
+        "test": "npm run build && jest"
+    },
+    "dependencies": {
+        "@types/bcryptjs": "^2.4.2",
+        "@types/configstore": "^4.0.0",
+        "@types/cookie-parser": "^1.4.2",
+        "@types/dockerode": "^2.5.34",
+        "@types/ejs": "^3.0.4",
+        "@types/express": "^4.17.7",
+        "@types/fs-extra": "^9.0.1",
+        "@types/http-proxy": "^1.17.4",
+        "@types/jest": "^26.0.10",
+        "@types/js-base64": "^3.0.0",
+        "@types/jsonwebtoken": "^8.5.0",
+        "@types/morgan": "^1.9.1",
+        "@types/multer": "^1.4.4",
+        "@types/on-finished": "^2.3.1",
+        "@types/request": "^2.48.5",
+        "@types/require-from-string": "^1.2.0",
+        "@types/serve-favicon": "^2.5.0",
+        "@types/ssh2": "^0.5.44",
+        "@types/tar": "^4.0.3",
+        "@types/uuid": "^8.3.0",
+        "@types/validator": "^13.1.0",
+        "axios": "^0.19.2",
+        "bcryptjs": "^2.4.3",
+        "body-parser": "^1.19.0",
+        "configstore": "^5.0.1",
+        "cookie-parser": "~1.4.5",
+        "debug": "~4.1.1",
+        "dockerode": "^3.2.1",
+        "ejs": "^3.1.3",
+        "express": "^4.17.1",
+        "fs-extra": "^9.0.1",
+        "http-proxy": "^1.18.1",
+        "is-valid-path": "^0.1.1",
+        "js-base64": "^3.4.5",
+        "jsonwebtoken": "^8.5.1",
+        "madge": "^3.9.2",
+        "moment": "^2.27.0",
+        "morgan": "^1.10.0",
+        "multer": "^1.4.2",
+        "on-finished": "^2.3.0",
+        "prettier": "^2.0.5",
+        "public-ip": "^4.0.2",
+        "recursive-readdir": "^2.2.2",
+        "request": "^2.88.2",
+        "require-from-string": "^2.0.2",
+        "serve-favicon": "~2.5.0",
+        "simple-git": "^2.17.0",
+        "ssh2": "^0.8.9",
+        "tar": "^6.0.5",
+        "tslint": "^6.1.3",
+        "typescript": "^3.9.7",
+        "uuid": "^8.3.0",
+        "validator": "^13.1.1",
+        "yaml": "^1.10.0"
+    },
+    "devDependencies": {
+        "jest": "^26.4.0",
+        "ts-jest": "^26.2.0"
+    }
 }
