@@ -1,13 +1,12 @@
-import uuid = require('uuid/v4')
-import ApiStatusCodes = require('../api/ApiStatusCodes')
-import CaptainConstants = require('../utils/CaptainConstants')
-import Logger = require('../utils/Logger')
-import configstore = require('configstore')
-import Authenticator = require('../user/Authenticator')
-import { CaptainEncryptor } from '../utils/Encryptor'
+import { v4 as uuid } from 'uuid'
+import ApiStatusCodes from '../api/ApiStatusCodes'
 import { IBuiltImage } from '../models/IBuiltImage'
-import Utils from '../utils/Utils'
+import Authenticator from '../user/Authenticator'
 import ApacheMd5 from '../utils/ApacheMd5'
+import CaptainConstants from '../utils/CaptainConstants'
+import CaptainEncryptor from '../utils/Encryptor'
+import Utils from '../utils/Utils'
+import configstore = require('configstore')
 
 const isValidPath = require('is-valid-path')
 
@@ -41,7 +40,7 @@ class AppsDataStore {
         const self = this
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 if (!appName) {
                     throw ApiStatusCodes.createError(
                         ApiStatusCodes.STATUS_ERROR_GENERIC,
@@ -132,7 +131,7 @@ class AppsDataStore {
                         if (!isValidPath(obj.containerPath)) {
                             throw ApiStatusCodes.createError(
                                 ApiStatusCodes.STATUS_ERROR_GENERIC,
-                                'Invalid containerPath: ' + obj.containerPath
+                                `Invalid containerPath: ${obj.containerPath}`
                             )
                         }
 
@@ -140,7 +139,7 @@ class AppsDataStore {
                             if (!isValidPath(obj.hostPath)) {
                                 throw ApiStatusCodes.createError(
                                     ApiStatusCodes.STATUS_ERROR_GENERIC,
-                                    'Invalid volume host path: ' + obj.hostPath
+                                    `Invalid volume host path: ${obj.hostPath}`
                                 )
                             }
                         } else {
@@ -150,14 +149,14 @@ class AppsDataStore {
                             ) {
                                 throw ApiStatusCodes.createError(
                                     ApiStatusCodes.STATUS_ERROR_GENERIC,
-                                    'Invalid volume name: ' + obj.volumeName
+                                    `Invalid volume name: ${obj.volumeName}`
                                 )
                             }
                         }
                     }
                 }
             })
-            .then(function() {
+            .then(function () {
                 let passwordToBeEncrypted = ''
                 let sshKeyToBeEncrypted = ''
                 let pushWebhook = app.appPushWebhook
@@ -194,8 +193,8 @@ class AppsDataStore {
 
                 return appToSave
             })
-            .then(function(appToSave: IAppDefSaved) {
-                self.data.set(APP_DEFINITIONS + '.' + appName, appToSave)
+            .then(function (appToSave: IAppDefSaved) {
+                self.data.set(`${APP_DEFINITIONS}.${appName}`, appToSave)
             })
     }
 
@@ -203,45 +202,73 @@ class AppsDataStore {
         if (!isNameAllowed(appName)) {
             throw ApiStatusCodes.createError(
                 ApiStatusCodes.STATUS_ERROR_BAD_NAME,
-                'App Name is not allow. Only lowercase letters and single hyphen is allow'
+                'App Name is not allowed. Only lowercase letters and single hyphens are allowed'
+            )
+        }
+
+        if (!!this.data.get(`${APP_DEFINITIONS}.${appName}`)) {
+            throw ApiStatusCodes.createError(
+                ApiStatusCodes.STATUS_ERROR_ALREADY_EXIST,
+                'App Name already exists. Please use a different name'
             )
         }
     }
 
-    renameApp(oldAppName: string, newAppName: string) {
+    renameApp(
+        authenticator: Authenticator,
+        oldAppName: string,
+        newAppName: string
+    ) {
         const self = this
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 self.nameAllowedOrThrow(newAppName)
                 return self.getAppDefinition(oldAppName)
             })
-            .then(function(appData) {
+            .then(function (appData) {
+                if (
+                    appData.appPushWebhook &&
+                    appData.appPushWebhook.pushWebhookToken
+                ) {
+                    const tokenVersion = uuid()
+                    return authenticator
+                        .getAppPushWebhookToken(newAppName, tokenVersion)
+                        .then((val) => {
+                            appData.appPushWebhook!.pushWebhookToken = val
+                            appData.appPushWebhook!.tokenVersion = tokenVersion
+                            return appData
+                        })
+                }
+
+                return appData
+            })
+            .then(function (appData) {
                 if (appData.appName) appData.appName = newAppName
                 appData.hasDefaultSubDomainSsl = false
-                self.data.delete(APP_DEFINITIONS + '.' + oldAppName)
-                self.saveApp(newAppName, appData)
+                return self.saveApp(newAppName, appData)
             })
-            .then(function() {
-                Utils.getDelayedPromise(2000)
+            .then(function () {
+                self.data.delete(`${APP_DEFINITIONS}.${oldAppName}`)
+                return Utils.getDelayedPromise(2000)
             })
     }
 
     getServiceName(appName: string) {
-        return 'srv-' + this.namepace + '--' + appName
+        return `srv-${this.namepace}--${appName}`
     }
 
     getVolumeName(volumeName: string) {
-        return this.namepace + '--' + volumeName
+        return `${this.namepace}--${volumeName}`
     }
 
     getAppDefinitions() {
         const self = this
-        return new Promise<IAllAppDefinitions>(function(resolve, reject) {
+        return new Promise<IAllAppDefinitions>(function (resolve, reject) {
             let allApps = self.data.get(APP_DEFINITIONS) || {}
             let allAppsUnencrypted: IAllAppDefinitions = {}
 
-            Object.keys(allApps).forEach(function(appName) {
+            Object.keys(allApps).forEach(function (appName) {
                 allAppsUnencrypted[appName] = allApps[appName]
                 const appUnencrypted = allAppsUnencrypted[appName]
 
@@ -282,9 +309,7 @@ class AppsDataStore {
     }
 
     getAppDefinition(appName: string) {
-        const self = this
-
-        return this.getAppDefinitions().then(function(allApps) {
+        return this.getAppDefinitions().then(function (allApps) {
             if (!appName) {
                 throw ApiStatusCodes.createError(
                     ApiStatusCodes.STATUS_ERROR_GENERIC,
@@ -308,7 +333,7 @@ class AppsDataStore {
     setSslForDefaultSubDomain(appName: string, isEnabled: boolean) {
         const self = this
 
-        return this.getAppDefinition(appName).then(function(app) {
+        return this.getAppDefinition(appName).then(function (app) {
             app.hasDefaultSubDomainSsl = !!isEnabled
             return self.saveApp(appName, app)
         })
@@ -317,20 +342,20 @@ class AppsDataStore {
     ensureAllAppsSubDomainSslDisabled() {
         const self = this
 
-        return this.getAppDefinitions().then(function(appDefinitions) {
+        return this.getAppDefinitions().then(function (appDefinitions) {
             const promises: (() => Promise<void>)[] = []
-            Object.keys(appDefinitions).forEach(appName => {
+            Object.keys(appDefinitions).forEach((appName) => {
                 const APP_NAME = appName
-                promises.push(function() {
+                promises.push(function () {
                     return Promise.resolve()
-                        .then(function() {
+                        .then(function () {
                             return self.getAppDefinition(APP_NAME)
                         })
-                        .then(function(app) {
+                        .then(function (app) {
                             app.forceSsl = false
                             return self.saveApp(APP_NAME, app)
                         })
-                        .then(function() {
+                        .then(function () {
                             return self.setSslForDefaultSubDomain(
                                 APP_NAME,
                                 false
@@ -345,7 +370,7 @@ class AppsDataStore {
     enableCustomDomainSsl(appName: string, customDomain: string) {
         const self = this
 
-        return self.getAppDefinition(appName).then(function(app) {
+        return self.getAppDefinition(appName).then(function (app) {
             app.customDomain = app.customDomain || []
 
             if (app.customDomain.length > 0) {
@@ -366,7 +391,7 @@ class AppsDataStore {
     removeCustomDomainForApp(appName: string, customDomain: string) {
         const self = this
 
-        return this.getAppDefinition(appName).then(function(app) {
+        return this.getAppDefinition(appName).then(function (app) {
             app.customDomain = app.customDomain || []
 
             const newDomains = []
@@ -394,7 +419,7 @@ class AppsDataStore {
     addCustomDomainForApp(appName: string, customDomain: string) {
         const self = this
 
-        return this.getAppDefinition(appName).then(function(app) {
+        return this.getAppDefinition(appName).then(function (app) {
             app.customDomain = app.customDomain || []
 
             if (app.customDomain.length > 0) {
@@ -425,7 +450,7 @@ class AppsDataStore {
         const self = this
 
         return this.getAppDefinition(appName) //
-            .then(function(app) {
+            .then(function (app) {
                 app.customDomain = app.customDomain || []
 
                 for (let idx = 0; idx < customDomains.length; idx++) {
@@ -444,7 +469,7 @@ class AppsDataStore {
     verifyCustomDomainBelongsToApp(appName: string, customDomain: string) {
         const self = this
 
-        return self.getAppDefinition(appName).then(function(app) {
+        return self.getAppDefinition(appName).then(function (app) {
             app.customDomain = app.customDomain || []
 
             if (app.customDomain.length > 0) {
@@ -484,7 +509,7 @@ class AppsDataStore {
         const self = this
 
         return this.getAppDefinition(appName) //
-            .then(function(app) {
+            .then(function (app) {
                 const versions = app.versions
 
                 let found = false
@@ -521,7 +546,7 @@ class AppsDataStore {
         }
         const self = this
 
-        return this.getAppDefinition(appName).then(function(app) {
+        return this.getAppDefinition(appName).then(function (app) {
             // Drop older versions
             app.versions = Utils.dropFirstElements(
                 app.versions,
@@ -551,7 +576,7 @@ class AppsDataStore {
                 timeStamp: new Date().toISOString(),
             })
 
-            return self.saveApp(appName, app).then(function() {
+            return self.saveApp(appName, app).then(function () {
                 return newVersionIndex
             })
         })
@@ -564,10 +589,10 @@ class AppsDataStore {
     ) {
         const self = this
         return Promise.resolve() //
-            .then(function() {
+            .then(function () {
                 return self.getAppDefinition(appName)
             })
-            .then(function(appLoaded) {
+            .then(function (appLoaded) {
                 appLoaded.deployedVersion = deployedVersion
                 appLoaded.versions = vers
                 return self.saveApp(appName, appLoaded)
@@ -591,16 +616,17 @@ class AppsDataStore {
         authenticator: Authenticator,
         customNginxConfig: string,
         preDeployFunction: string,
+        serviceUpdateOverride: string,
         websocketSupport: boolean
     ) {
         const self = this
         let appObj: IAppDef
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 return self.getAppDefinition(appName)
             })
-            .then(function(appLoaded) {
+            .then(function (appLoaded) {
                 appObj = appLoaded
 
                 if (
@@ -636,7 +662,7 @@ class AppsDataStore {
                             appName,
                             appObj.appPushWebhook.tokenVersion
                         )
-                        .then(function(val) {
+                        .then(function (val) {
                             appObj.appPushWebhook!.pushWebhookToken = val
                         })
                 } else {
@@ -644,7 +670,7 @@ class AppsDataStore {
                     return Promise.resolve(undefined)
                 }
             })
-            .then(function() {
+            .then(function () {
                 instanceCount = Number(instanceCount)
 
                 if (instanceCount >= 0) {
@@ -663,6 +689,7 @@ class AppsDataStore {
                 appObj.nodeId = nodeId
                 appObj.customNginxConfig = customNginxConfig
                 appObj.preDeployFunction = preDeployFunction
+                appObj.serviceUpdateOverride = serviceUpdateOverride
                 appObj.description = description
 
                 if (httpAuth && httpAuth.user) {
@@ -749,7 +776,7 @@ class AppsDataStore {
                     }
                 }
             })
-            .then(function() {
+            .then(function () {
                 return self.saveApp(appName, appObj)
             })
     }
@@ -757,18 +784,18 @@ class AppsDataStore {
     deleteAppDefinition(appName: string) {
         const self = this
 
-        return new Promise<void>(function(resolve, reject) {
+        return new Promise<void>(function (resolve, reject) {
             if (!isNameAllowed(appName)) {
                 reject(
                     ApiStatusCodes.createError(
                         ApiStatusCodes.STATUS_ERROR_BAD_NAME,
-                        'App Name is not allow. Only lowercase letters and single hyphen is allow'
+                        'App Name is not allowed. Only lowercase letters and single hyphens are allowed'
                     )
                 )
                 return
             }
 
-            if (!self.data.get(APP_DEFINITIONS + '.' + appName)) {
+            if (!self.data.get(`${APP_DEFINITIONS}.${appName}`)) {
                 reject(
                     ApiStatusCodes.createError(
                         ApiStatusCodes.STATUS_ERROR_GENERIC,
@@ -778,10 +805,10 @@ class AppsDataStore {
                 return
             }
 
-            self.data.delete(APP_DEFINITIONS + '.' + appName)
+            self.data.delete(`${APP_DEFINITIONS}.${appName}`)
             resolve()
-        }).then(function() {
-            Utils.getDelayedPromise(2000)
+        }).then(function () {
+            return Utils.getDelayedPromise(2000)
         })
     }
 
@@ -795,18 +822,18 @@ class AppsDataStore {
     registerAppDefinition(appName: string, hasPersistentData: boolean) {
         const self = this
 
-        return new Promise<IAppDef>(function(resolve, reject) {
+        return new Promise<IAppDef>(function (resolve, reject) {
             if (!isNameAllowed(appName)) {
                 reject(
                     ApiStatusCodes.createError(
                         ApiStatusCodes.STATUS_ERROR_BAD_NAME,
-                        'App Name is not allow. Only lowercase letters and single hyphen is allow'
+                        'App Name is not allowed. Only lowercase letters and single hyphens are allowed'
                     )
                 )
                 return
             }
 
-            if (!!self.data.get(APP_DEFINITIONS + '.' + appName)) {
+            if (!!self.data.get(`${APP_DEFINITIONS}.${appName}`)) {
                 reject(
                     ApiStatusCodes.createError(
                         ApiStatusCodes.STATUS_ERROR_ALREADY_EXIST,
@@ -836,10 +863,10 @@ class AppsDataStore {
             }
 
             resolve(defaultAppDefinition)
-        }).then(function(app) {
+        }).then(function (app) {
             return self.saveApp(appName, app)
         })
     }
 }
 
-export = AppsDataStore
+export default AppsDataStore
