@@ -2,12 +2,12 @@
  * Created by kasra on 27/06/17.
  */
 
-import DataStore = require('./DataStore')
+import DataStore from './DataStore'
 
 const dataStoreCache: IHashMapGeneric<DataStore> = {}
 
-export = {
-    getDataStore: function(namespace: string) {
+export default {
+    getDataStore: function (namespace: string) {
         if (!namespace) {
             throw new Error('NameSpace is empty')
         }
