@@ -1,16 +1,16 @@
-import DataStoreProvider = require('../datastore/DataStoreProvider')
-import Authenticator = require('../user/Authenticator')
-import CaptainConstants = require('../utils/CaptainConstants')
-import CaptainManager = require('../user/system/CaptainManager')
-import ServiceManager = require('../user/ServiceManager')
+import { NextFunction, Request, Response } from 'express'
+import ApiStatusCodes from '../api/ApiStatusCodes'
+import BaseApi from '../api/BaseApi'
+import DataStoreProvider from '../datastore/DataStoreProvider'
 import DockerApiProvider from '../docker/DockerApi'
-import BaseApi = require('../api/BaseApi')
-import UserModel = require('../models/InjectionInterfaces')
-import Logger = require('../utils/Logger')
-import { Response, Request, NextFunction } from 'express'
+import * as UserModel from '../models/InjectionInterfaces'
 import { CaptainError } from '../models/OtherTypes'
-import InjectionExtractor = require('./InjectionExtractor')
-import ApiStatusCodes = require('../api/ApiStatusCodes')
+import Authenticator from '../user/Authenticator'
+import ServiceManager from '../user/ServiceManager'
+import CaptainManager from '../user/system/CaptainManager'
+import CaptainConstants from '../utils/CaptainConstants'
+import Logger from '../utils/Logger'
+import InjectionExtractor from './InjectionExtractor'
 
 const dockerApi = DockerApiProvider.get()
 
@@ -18,7 +18,7 @@ const dockerApi = DockerApiProvider.get()
  * Global dependency injection module
  */
 export function injectGlobal() {
-    return function(req: Request, res: Response, next: NextFunction) {
+    return function (req: Request, res: Response, next: NextFunction) {
         const locals = res.locals
 
         locals.initialized = CaptainManager.get().isInitialized()
@@ -43,7 +43,7 @@ export function injectGlobal() {
  * User dependency injection module
  */
 export function injectUser() {
-    return function(req: Request, res: Response, next: NextFunction) {
+    return function (req: Request, res: Response, next: NextFunction) {
         if (InjectionExtractor.extractUserFromInjected(res).user) {
             next()
             return // user is already injected by another layer
@@ -53,7 +53,7 @@ export function injectUser() {
 
         Authenticator.getAuthenticator(namespace)
             .decodeAuthToken(req.header(CaptainConstants.headerAuth) || '')
-            .then(function(userDecoded) {
+            .then(function (userDecoded) {
                 if (userDecoded) {
                     const datastore = DataStoreProvider.getDataStore(namespace)
 
@@ -76,7 +76,7 @@ export function injectUser() {
 
                 next()
             })
-            .catch(function(error: CaptainError) {
+            .catch(function (error: CaptainError) {
                 if (error && error.captainErrorType) {
                     res.send(
                         new BaseApi(error.captainErrorType, error.apiMessage)
@@ -94,9 +94,9 @@ export function injectUser() {
  * A pseudo user injection. Only used for webhooks. Can only trigger certain actions.
  */
 export function injectUserForWebhook() {
-    return function(req: Request, res: Response, next: NextFunction) {
-        const token = req.query.token
-        const namespace = req.query.namespace
+    return function (req: Request, res: Response, next: NextFunction) {
+        const token = req.query.token as string
+        const namespace = req.query.namespace as string
         let app = undefined
 
         if (!token || !namespace) {
@@ -111,14 +111,14 @@ export function injectUserForWebhook() {
 
         Authenticator.getAuthenticator(namespace)
             .decodeAppPushWebhookToken(token)
-            .then(function(data) {
+            .then(function (data) {
                 decodedInfo = data
 
                 return dataStore
                     .getAppsDataStore()
                     .getAppDefinition(decodedInfo.appName)
             })
-            .then(function(appFound) {
+            .then(function (appFound) {
                 app = appFound
 
                 if (
@@ -152,7 +152,7 @@ export function injectUserForWebhook() {
 
                 next()
             })
-            .catch(function(error) {
+            .catch(function (error) {
                 Logger.e(error)
                 res.locals.app = undefined
                 next()
@@ -165,17 +165,17 @@ export function injectUserForWebhook() {
  * this is the only way that we can secure the call
  */
 export function injectUserUsingCookieDataOnly() {
-    return function(req: Request, res: Response, next: NextFunction) {
+    return function (req: Request, res: Response, next: NextFunction) {
         Authenticator.getAuthenticator(CaptainConstants.rootNameSpace)
             .decodeAuthTokenFromCookies(
                 req.cookies[CaptainConstants.headerCookieAuth]
             )
-            .then(function(user) {
+            .then(function (user) {
                 res.locals.user = user
 
                 next()
             })
-            .catch(function(error) {
+            .catch(function (error) {
                 if (error && error.captainErrorType) {
                     res.send(
                         new BaseApi(error.captainErrorType, error.apiMessage)
