@@ -1,17 +1,16 @@
 import express = require('express')
-import BaseApi = require('../../api/BaseApi')
-import ApiStatusCodes = require('../../api/ApiStatusCodes')
-import Injector = require('../../injection/Injector')
-import SystemRouter = require('./system/SystemRouter')
-import AppsRouter = require('./apps/AppsRouter')
-import Logger = require('../../utils/Logger')
-import RegistriesRouter = require('./registeries/RegistriesRouter')
-import onFinished = require('on-finished')
-import InjectionExtractor = require('../../injection/InjectionExtractor')
-import CaptainManager = require('../../user/system/CaptainManager')
+import ApiStatusCodes from '../../api/ApiStatusCodes'
+import BaseApi from '../../api/BaseApi'
+import InjectionExtractor from '../../injection/InjectionExtractor'
+import * as Injector from '../../injection/Injector'
+import Authenticator from '../../user/Authenticator'
+import EnvVars from '../../utils/EnvVars'
 import Utils from '../../utils/Utils'
-import EnvVars = require('../../utils/EnvVars')
-import Authenticator = require('../../user/Authenticator')
+import AppsRouter from './apps/AppsRouter'
+import OneClickAppRouter from './oneclick/OneClickAppRouter'
+import RegistriesRouter from './registeries/RegistriesRouter'
+import SystemRouter from './system/SystemRouter'
+import onFinished = require('on-finished')
 
 const router = express.Router()
 
@@ -21,7 +20,7 @@ router.use('/apps/webhooks/', Injector.injectUserForWebhook())
 
 router.use(Injector.injectUser())
 
-router.use(function(req, res, next) {
+router.use(function (req, res, next) {
     const user = InjectionExtractor.extractUserFromInjected(res).user
 
     if (!user) {
@@ -53,18 +52,26 @@ router.use(function(req, res, next) {
         return
     }
 
-    const serviceManager = user.serviceManager
-
     // All requests except GET might be making changes to some stuff that are not designed for an asynchronous process
     // I'm being extra cautious. But removal of this lock mechanism requires testing and consideration of edge cases.
     if (Utils.isNotGetRequest(req)) {
-        if (EnvVars.IS_DEMO_MODE) {
-            let response = new BaseApi(
-                ApiStatusCodes.STATUS_ERROR_GENERIC,
-                'Demo mode is only for viewing purposes.'
-            )
-            res.send(response)
-            return
+        if (!!EnvVars.DEMO_MODE_ADMIN_IP) {
+            const realIp = `${req.headers['x-real-ip']}`
+            const forwardedIp = `${req.headers['x-forwarded-for']}`
+            if (
+                !realIp ||
+                !Utils.isValidIp(realIp) ||
+                realIp !== forwardedIp ||
+                EnvVars.DEMO_MODE_ADMIN_IP !== realIp
+            ) {
+                let response = new BaseApi(
+                    ApiStatusCodes.STATUS_ERROR_GENERIC,
+                    'Demo mode is only for viewing purposes.'
+                )
+                res.send(response)
+
+                return
+            }
         }
 
         if (threadLockNamespace[namespace]) {
@@ -76,7 +83,7 @@ router.use(function(req, res, next) {
 
         // we don't want the same space to go under two simultaneous changes
         threadLockNamespace[namespace] = true
-        onFinished(res, function() {
+        onFinished(res, function () {
             threadLockNamespace[namespace] = false
         })
     }
@@ -84,27 +91,27 @@ router.use(function(req, res, next) {
     next()
 })
 
-router.post('/changepassword/', function(req, res, next) {
+router.post('/changepassword/', function (req, res, next) {
     const namespace = InjectionExtractor.extractUserFromInjected(res).user
         .namespace
     const dataStore = InjectionExtractor.extractUserFromInjected(res).user
         .dataStore
 
     Promise.resolve() //
-        .then(function(data) {
+        .then(function (data) {
             return dataStore.getHashedPassword()
         })
-        .then(function(savedHashedPassword) {
+        .then(function (savedHashedPassword) {
             return Authenticator.getAuthenticator(namespace).changepass(
                 req.body.oldPassword,
                 req.body.newPassword,
                 savedHashedPassword
             )
         })
-        .then(function(hashedPassword) {
+        .then(function (hashedPassword) {
             return dataStore.setHashedPassword(hashedPassword)
         })
-        .then(function() {
+        .then(function () {
             res.send(new BaseApi(ApiStatusCodes.STATUS_OK, 'Password changed.'))
         })
         .catch(ApiStatusCodes.createCatcher(res))
@@ -112,8 +119,10 @@ router.post('/changepassword/', function(req, res, next) {
 
 router.use('/apps/', AppsRouter)
 
+router.use('/oneclick/', OneClickAppRouter)
+
 router.use('/registries/', RegistriesRouter)
 
 router.use('/system/', SystemRouter)
 
-export = router
+export default router
