@@ -1,29 +1,29 @@
 import express = require('express')
-import BaseApi = require('../../../api/BaseApi')
-import ApiStatusCodes = require('../../../api/ApiStatusCodes')
-import Logger = require('../../../utils/Logger')
-import CaptainManager = require('../../../user/system/CaptainManager')
-import Validator = require('validator')
-import SystemRouteSelfHostRegistry = require('./selfhostregistry/SystemRouteSelfHostRegistry')
-import CaptainConstants = require('../../../utils/CaptainConstants')
-import InjectionExtractor = require('../../../injection/InjectionExtractor')
-import Utils from '../../../utils/Utils'
-import * as path from 'path'
-import DockerUtils from '../../../docker/DockerUtils'
+import validator from 'validator'
+import ApiStatusCodes from '../../../api/ApiStatusCodes'
+import BaseApi from '../../../api/BaseApi'
 import DockerApi from '../../../docker/DockerApi'
+import DockerUtils from '../../../docker/DockerUtils'
+import InjectionExtractor from '../../../injection/InjectionExtractor'
+import CaptainManager from '../../../user/system/CaptainManager'
+import VersionManager from '../../../user/system/VersionManager'
+import CaptainConstants from '../../../utils/CaptainConstants'
+import Logger from '../../../utils/Logger'
+import Utils from '../../../utils/Utils'
+import SystemRouteSelfHostRegistry from './selfhostregistry/SystemRouteSelfHostRegistry'
 
 const router = express.Router()
 
 router.use('/selfhostregistry/', SystemRouteSelfHostRegistry)
 
-router.post('/createbackup/', function(req, res, next) {
+router.post('/createbackup/', function (req, res, next) {
     const backupManager = CaptainManager.get().getBackupManager()
 
     Promise.resolve()
-        .then(function() {
+        .then(function () {
             return backupManager.createBackup(CaptainManager.get())
         })
-        .then(function(backupInfo) {
+        .then(function (backupInfo) {
             const baseApi = new BaseApi(
                 ApiStatusCodes.STATUS_OK,
                 'Backup created.'
@@ -34,7 +34,7 @@ router.post('/createbackup/', function(req, res, next) {
         .catch(ApiStatusCodes.createCatcher(res))
 })
 
-router.post('/changerootdomain/', function(req, res, next) {
+router.post('/changerootdomain/', function (req, res, next) {
     let requestedCustomDomain = Utils.removeHttpHttps(
         (req.body.rootDomain || '').toLowerCase()
     )
@@ -56,7 +56,7 @@ router.post('/changerootdomain/', function(req, res, next) {
 
     CaptainManager.get()
         .changeCaptainRootDomain(requestedCustomDomain, !!req.body.force)
-        .then(function() {
+        .then(function () {
             res.send(
                 new BaseApi(ApiStatusCodes.STATUS_OK, 'Root domain changed.')
             )
@@ -64,7 +64,7 @@ router.post('/changerootdomain/', function(req, res, next) {
         .catch(ApiStatusCodes.createCatcher(res))
 })
 
-router.post('/enablessl/', function(req, res, next) {
+router.post('/enablessl/', function (req, res, next) {
     const emailAddress = req.body.emailAddress || ''
 
     if (
@@ -75,7 +75,7 @@ router.post('/enablessl/', function(req, res, next) {
         emailAddress.indexOf('%') >= 0 ||
         emailAddress.indexOf(' ') >= 0 ||
         emailAddress.indexOf('\\') >= 0 ||
-        !Validator.isEmail(emailAddress)
+        !validator.isEmail(emailAddress)
     ) {
         res.send(
             new BaseApi(
@@ -88,44 +88,45 @@ router.post('/enablessl/', function(req, res, next) {
 
     CaptainManager.get()
         .enableSsl(emailAddress)
-        .then(function() {
+        .then(function () {
             // This is necessary as the CLI immediately tries to connect to https://captain.root.com
             // Without this delay it'll fail to connect
             Logger.d('Waiting for 7 seconds...')
             return Utils.getDelayedPromise(7000)
         })
-        .then(function() {
+        .then(function () {
             res.send(new BaseApi(ApiStatusCodes.STATUS_OK, 'Root SSL Enabled.'))
         })
         .catch(ApiStatusCodes.createCatcher(res))
 })
 
-router.post('/forcessl/', function(req, res, next) {
+router.post('/forcessl/', function (req, res, next) {
     let isEnabled = !!req.body.isEnabled
 
     CaptainManager.get()
         .forceSsl(isEnabled)
-        .then(function() {
+        .then(function () {
             res.send(
                 new BaseApi(
                     ApiStatusCodes.STATUS_OK,
-                    'Non-SSL traffic is now ' +
-                        (isEnabled ? 'rejected.' : 'allowed.')
+                    `Non-SSL traffic is now ${
+                        isEnabled ? 'rejected.' : 'allowed.'
+                    }`
                 )
             )
         })
         .catch(ApiStatusCodes.createCatcher(res))
 })
 
-router.get('/info/', function(req, res, next) {
+router.get('/info/', function (req, res, next) {
     const dataStore = InjectionExtractor.extractUserFromInjected(res).user
         .dataStore
 
     return Promise.resolve()
-        .then(function() {
+        .then(function () {
             return dataStore.getHasRootSsl()
         })
-        .then(function(hasRootSsl) {
+        .then(function (hasRootSsl) {
             return {
                 hasRootSsl: hasRootSsl,
                 forceSsl: CaptainManager.get().getForceSslValue(),
@@ -134,7 +135,7 @@ router.get('/info/', function(req, res, next) {
                     : '',
             }
         })
-        .then(function(data) {
+        .then(function (data) {
             let baseApi = new BaseApi(
                 ApiStatusCodes.STATUS_OK,
                 'Captain info retrieved'
@@ -145,14 +146,12 @@ router.get('/info/', function(req, res, next) {
         .catch(ApiStatusCodes.createCatcher(res))
 })
 
-router.get('/loadbalancerinfo/', function(req, res, next) {
+router.get('/loadbalancerinfo/', function (req, res, next) {
     return Promise.resolve()
-        .then(function() {
-            return CaptainManager.get()
-                .getLoadBalanceManager()
-                .getInfo()
+        .then(function () {
+            return CaptainManager.get().getLoadBalanceManager().getInfo()
         })
-        .then(function(data) {
+        .then(function (data) {
             let baseApi = new BaseApi(
                 ApiStatusCodes.STATUS_OK,
                 'Load Balancer info retrieved'
@@ -163,69 +162,30 @@ router.get('/loadbalancerinfo/', function(req, res, next) {
         .catch(ApiStatusCodes.createCatcher(res))
 })
 
-router.get('/versionInfo/', function(req, res, next) {
+router.get('/versionInfo/', function (req, res, next) {
     return Promise.resolve()
-        .then(function() {
-            return CaptainManager.get().getCaptainImageTags()
+        .then(function () {
+            return VersionManager.get().getCaptainImageTags()
         })
-        .then(function(tagList) {
+        .then(function (data) {
             let baseApi = new BaseApi(
                 ApiStatusCodes.STATUS_OK,
                 'Version Info Retrieved'
             )
-            let currentVersion = CaptainConstants.configs.version.split('.')
-            let latestVersion = CaptainConstants.configs.version.split('.')
-
-            let canUpdate = false
-
-            for (let i = 0; i < tagList.length; i++) {
-                let tag = tagList[i].split('.')
-
-                if (tag.length !== 3) {
-                    continue
-                }
-
-                if (Number(tag[0]) > Number(currentVersion[0])) {
-                    canUpdate = true
-                    latestVersion = tag
-                    break
-                } else if (
-                    Number(tag[0]) === Number(currentVersion[0]) &&
-                    Number(tag[1]) > Number(currentVersion[1])
-                ) {
-                    canUpdate = true
-                    latestVersion = tag
-                    break
-                } else if (
-                    Number(tag[0]) === Number(currentVersion[0]) &&
-                    Number(tag[1]) === Number(currentVersion[1]) &&
-                    Number(tag[2]) > Number(currentVersion[2])
-                ) {
-                    canUpdate = true
-                    latestVersion = tag
-                    break
-                }
-            }
-
-            baseApi.data = {
-                currentVersion: currentVersion.join('.'),
-                latestVersion: latestVersion.join('.'),
-                canUpdate: canUpdate,
-            }
-
+            baseApi.data = data
             res.send(baseApi)
         })
         .catch(ApiStatusCodes.createCatcher(res))
 })
 
-router.post('/versionInfo/', function(req, res, next) {
+router.post('/versionInfo/', function (req, res, next) {
     let latestVersion = req.body.latestVersion
 
     return Promise.resolve()
-        .then(function() {
-            return CaptainManager.get().updateCaptain(latestVersion)
+        .then(function () {
+            return VersionManager.get().updateCaptain(latestVersion)
         })
-        .then(function() {
+        .then(function () {
             let baseApi = new BaseApi(
                 ApiStatusCodes.STATUS_OK,
                 'Captain update process has started...'
@@ -235,20 +195,20 @@ router.post('/versionInfo/', function(req, res, next) {
         .catch(ApiStatusCodes.createCatcher(res))
 })
 
-router.get('/netdata/', function(req, res, next) {
+router.get('/netdata/', function (req, res, next) {
     const dataStore = InjectionExtractor.extractUserFromInjected(res).user
         .dataStore
 
     return Promise.resolve()
-        .then(function() {
+        .then(function () {
             return dataStore.getNetDataInfo()
         })
-        .then(function(data) {
-            data.netDataUrl =
-                CaptainConstants.captainSubDomain +
-                '.' +
-                dataStore.getRootDomain() +
+        .then(function (data) {
+            data.netDataUrl = `${
+                CaptainConstants.captainSubDomain
+            }.${dataStore.getRootDomain()}${
                 CaptainConstants.netDataRelativePath
+            }`
             let baseApi = new BaseApi(
                 ApiStatusCodes.STATUS_OK,
                 'Netdata info retrieved'
@@ -259,16 +219,16 @@ router.get('/netdata/', function(req, res, next) {
         .catch(ApiStatusCodes.createCatcher(res))
 })
 
-router.post('/netdata/', function(req, res, next) {
+router.post('/netdata/', function (req, res, next) {
     let netDataInfo = req.body.netDataInfo
     netDataInfo.netDataUrl = undefined // Frontend app returns this value, but we really don't wanna save this.
     // root address is subject to change.
 
     return Promise.resolve()
-        .then(function() {
+        .then(function () {
             return CaptainManager.get().updateNetDataInfo(netDataInfo)
         })
-        .then(function() {
+        .then(function () {
             let baseApi = new BaseApi(
                 ApiStatusCodes.STATUS_OK,
                 'Netdata info is updated'
@@ -278,12 +238,12 @@ router.post('/netdata/', function(req, res, next) {
         .catch(ApiStatusCodes.createCatcher(res))
 })
 
-router.get('/nginxconfig/', function(req, res, next) {
+router.get('/nginxconfig/', function (req, res, next) {
     return Promise.resolve()
-        .then(function() {
+        .then(function () {
             return CaptainManager.get().getNginxConfig()
         })
-        .then(function(data) {
+        .then(function (data) {
             let baseApi = new BaseApi(
                 ApiStatusCodes.STATUS_OK,
                 'Nginx config retrieved'
@@ -294,18 +254,18 @@ router.get('/nginxconfig/', function(req, res, next) {
         .catch(ApiStatusCodes.createCatcher(res))
 })
 
-router.post('/nginxconfig/', function(req, res, next) {
+router.post('/nginxconfig/', function (req, res, next) {
     let baseConfigCustomValue = req.body.baseConfig.customValue
     let captainConfigCustomValue = req.body.captainConfig.customValue
 
     return Promise.resolve()
-        .then(function() {
+        .then(function () {
             return CaptainManager.get().setNginxConfig(
                 baseConfigCustomValue,
                 captainConfigCustomValue
             )
         })
-        .then(function() {
+        .then(function () {
             let baseApi = new BaseApi(
                 ApiStatusCodes.STATUS_OK,
                 'Nginx config is updated'
@@ -315,12 +275,12 @@ router.post('/nginxconfig/', function(req, res, next) {
         .catch(ApiStatusCodes.createCatcher(res))
 })
 
-router.get('/nodes/', function(req, res, next) {
+router.get('/nodes/', function (req, res, next) {
     return Promise.resolve()
-        .then(function() {
+        .then(function () {
             return CaptainManager.get().getNodesInfo()
         })
-        .then(function(data) {
+        .then(function (data) {
             let baseApi = new BaseApi(
                 ApiStatusCodes.STATUS_OK,
                 'Node info retrieved'
@@ -331,7 +291,7 @@ router.get('/nodes/', function(req, res, next) {
         .catch(ApiStatusCodes.createCatcher(res))
 })
 
-router.post('/nodes/', function(req, res, next) {
+router.post('/nodes/', function (req, res, next) {
     const MANAGER = 'manager'
     const WORKER = 'worker'
     const registryHelper = InjectionExtractor.extractUserFromInjected(
@@ -357,6 +317,8 @@ router.post('/nodes/', function(req, res, next) {
     let privateKey = req.body.privateKey
     let remoteNodeIpAddress = req.body.remoteNodeIpAddress
     let captainIpAddress = req.body.captainIpAddress
+    let sshPort = parseInt(req.body.sshPort) || 22
+    let sshUser = (req.body.sshUser || 'root').trim()
 
     if (!captainIpAddress || !remoteNodeIpAddress || !privateKey) {
         res.send(
@@ -369,10 +331,10 @@ router.post('/nodes/', function(req, res, next) {
     }
 
     return Promise.resolve()
-        .then(function() {
+        .then(function () {
             return registryHelper.getDefaultPushRegistryId()
         })
-        .then(function(defaultRegistry) {
+        .then(function (defaultRegistry) {
             if (!defaultRegistry) {
                 throw ApiStatusCodes.createError(
                     ApiStatusCodes.STATUS_ERROR_GENERIC,
@@ -380,16 +342,18 @@ router.post('/nodes/', function(req, res, next) {
                 )
             }
         })
-        .then(function() {
+        .then(function () {
             return DockerUtils.joinDockerNode(
                 DockerApi.get(),
+                sshUser,
+                sshPort,
                 captainIpAddress,
                 isManager,
                 remoteNodeIpAddress,
                 privateKey
             )
         })
-        .then(function() {
+        .then(function () {
             let msg = 'Docker node is successfully joined.'
             Logger.d(msg)
             res.send(new BaseApi(ApiStatusCodes.STATUS_OK, msg))
@@ -397,4 +361,4 @@ router.post('/nodes/', function(req, res, next) {
         .catch(ApiStatusCodes.createCatcher(res))
 })
 
-export = router
+export default router
