@@ -1,18 +1,17 @@
 import express = require('express')
-import BaseApi = require('../../../../api/BaseApi')
-import ApiStatusCodes = require('../../../../api/ApiStatusCodes')
-import Logger = require('../../../../utils/Logger')
-import CaptainManager = require('../../../../user/system/CaptainManager')
-import Validator = require('validator')
-import CaptainConstants = require('../../../../utils/CaptainConstants')
-import InjectionExtractor = require('../../../../injection/InjectionExtractor')
-import uuid = require('uuid/v4')
+import { v4 as uuid } from 'uuid'
+import ApiStatusCodes from '../../../../api/ApiStatusCodes'
+import BaseApi from '../../../../api/BaseApi'
+import InjectionExtractor from '../../../../injection/InjectionExtractor'
 import { IRegistryTypes } from '../../../../models/IRegistryInfo'
+import CaptainManager from '../../../../user/system/CaptainManager'
+import CaptainConstants from '../../../../utils/CaptainConstants'
+import Logger from '../../../../utils/Logger'
 
 const router = express.Router()
 
 // ERRORS if a local already exists in DB
-router.post('/enableregistry/', function(req, res, next) {
+router.post('/enableregistry/', function (req, res, next) {
     const captainManager = CaptainManager.get()
     const password = uuid()
     const registryHelper = InjectionExtractor.extractUserFromInjected(
@@ -20,20 +19,18 @@ router.post('/enableregistry/', function(req, res, next) {
     ).user.serviceManager.getRegistryHelper()
 
     return Promise.resolve()
-        .then(function() {
-            return CaptainManager.get()
-                .getDockerRegistry()
-                .enableRegistrySsl()
+        .then(function () {
+            return CaptainManager.get().getDockerRegistry().enableRegistrySsl()
         })
-        .then(function() {
+        .then(function () {
             return captainManager
                 .getDockerRegistry()
                 .ensureDockerRegistryRunningOnThisNode(password)
         })
-        .then(function() {
+        .then(function () {
             return registryHelper.getAllRegistries()
         })
-        .then(function(allRegs) {
+        .then(function (allRegs) {
             for (let index = 0; index < allRegs.length; index++) {
                 const element = allRegs[index]
                 if (element.registryType === IRegistryTypes.LOCAL_REG) {
@@ -56,7 +53,7 @@ router.post('/enableregistry/', function(req, res, next) {
                 IRegistryTypes.LOCAL_REG
             )
         })
-        .then(function() {
+        .then(function () {
             let msg = 'Local registry is created.'
             Logger.d(msg)
             res.send(new BaseApi(ApiStatusCodes.STATUS_OK, msg))
@@ -65,17 +62,17 @@ router.post('/enableregistry/', function(req, res, next) {
 })
 
 // ERRORS if default push is this
-router.post('/disableregistry/', function(req, res, next) {
+router.post('/disableregistry/', function (req, res, next) {
     const captainManager = CaptainManager.get()
     const registryHelper = InjectionExtractor.extractUserFromInjected(
         res
     ).user.serviceManager.getRegistryHelper()
 
     return Promise.resolve()
-        .then(function() {
+        .then(function () {
             return registryHelper.getAllRegistries()
         })
-        .then(function(regs) {
+        .then(function (regs) {
             let localRegistryId = ''
             for (let idx = 0; idx < regs.length; idx++) {
                 const element = regs[idx]
@@ -86,10 +83,10 @@ router.post('/disableregistry/', function(req, res, next) {
 
             return registryHelper.deleteRegistry(localRegistryId, true)
         })
-        .then(function() {
+        .then(function () {
             return captainManager.getDockerRegistry().ensureServiceRemoved()
         })
-        .then(function() {
+        .then(function () {
             let msg = 'Local registry is removed.'
             Logger.d(msg)
             res.send(new BaseApi(ApiStatusCodes.STATUS_OK, msg))
@@ -97,4 +94,4 @@ router.post('/disableregistry/', function(req, res, next) {
         .catch(ApiStatusCodes.createCatcher(res))
 })
 
-export = router
+export default router
