@@ -1,12 +1,11 @@
 import jwt = require('jsonwebtoken')
-import uuid = require('uuid/v4')
-import bcrypt = require('bcryptjs')
-
-import ApiStatusCodes = require('../api/ApiStatusCodes')
-import EnvVar = require('../utils/EnvVars')
-import CaptainConstants = require('../utils/CaptainConstants')
-import Logger = require('../utils/Logger')
+import { v4 as uuid } from 'uuid'
+import ApiStatusCodes from '../api/ApiStatusCodes'
 import { UserJwt } from '../models/UserJwt'
+import CaptainConstants from '../utils/CaptainConstants'
+import EnvVar from '../utils/EnvVars'
+import Logger from '../utils/Logger'
+import bcrypt = require('bcryptjs')
 
 const captainDefaultPassword = EnvVar.DEFAULT_PASSWORD || 'captain42'
 
@@ -32,7 +31,7 @@ class Authenticator {
         newPass = newPass || ''
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 if (!oldPass || !newPass || newPass.length < 8) {
                     throw ApiStatusCodes.createError(
                         ApiStatusCodes.STATUS_ERROR_GENERIC,
@@ -42,7 +41,7 @@ class Authenticator {
 
                 return self.isPasswordCorrect(oldPass, savedHashedPassword)
             })
-            .then(function(isPasswordCorrect) {
+            .then(function (isPasswordCorrect) {
                 if (!isPasswordCorrect) {
                     throw ApiStatusCodes.createError(
                         ApiStatusCodes.STATUS_WRONG_PASSWORD,
@@ -64,7 +63,7 @@ class Authenticator {
     isPasswordCorrect(password: string, savedHashedPassword: string) {
         const self = this
 
-        return Promise.resolve().then(function() {
+        return Promise.resolve().then(function () {
             password = password || ''
 
             if (!savedHashedPassword) {
@@ -94,10 +93,10 @@ class Authenticator {
         const self = this
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 return self.isPasswordCorrect(password, savedHashedPassword)
             })
-            .then(function(isPasswordCorrect) {
+            .then(function (isPasswordCorrect) {
                 if (!isPasswordCorrect) {
                     throw ApiStatusCodes.createError(
                         ApiStatusCodes.STATUS_WRONG_PASSWORD,
@@ -127,11 +126,11 @@ class Authenticator {
     decodeAuthToken(token: string, keySuffix?: string) {
         const self = this
 
-        return new Promise<UserJwt>(function(resolve, reject) {
+        return new Promise<UserJwt>(function (resolve, reject) {
             jwt.verify(
                 token,
                 self.encryptionKey + (keySuffix ? keySuffix : ''),
-                function(err, rawDecoded: { data: UserJwt }) {
+                function (err, rawDecoded: { data: UserJwt }) {
                     if (err) {
                         Logger.e(err)
                         reject(
@@ -225,7 +224,7 @@ class Authenticator {
         const self = this
         obj.namespace = self.namespace
 
-        return Promise.resolve().then(function() {
+        return Promise.resolve().then(function () {
             return jwt.sign(
                 {
                     data: obj,
@@ -243,11 +242,11 @@ class Authenticator {
     decodeGenericToken(token: string, keySuffix: string) {
         const self = this
 
-        return new Promise<any>(function(resolve, reject) {
+        return new Promise<any>(function (resolve, reject) {
             jwt.verify(
                 token,
                 self.encryptionKey + (keySuffix ? keySuffix : ''),
-                function(err, rawDecoded: { data: any }) {
+                function (err, rawDecoded: { data: any }) {
                     if (err) {
                         Logger.e(err)
                         reject(
@@ -309,4 +308,4 @@ class Authenticator {
     }
 }
 
-export = Authenticator
+export default Authenticator
