@@ -1,23 +1,22 @@
-import uuid = require('uuid/v4')
+import { v4 as uuid } from 'uuid'
+import ApiStatusCodes from '../../api/ApiStatusCodes'
+import DataStore from '../../datastore/DataStore'
+import DataStoreProvider from '../../datastore/DataStoreProvider'
+import DockerApi from '../../docker/DockerApi'
+import { IRegistryInfo, IRegistryTypes } from '../../models/IRegistryInfo'
+import CaptainConstants from '../../utils/CaptainConstants'
+import Logger from '../../utils/Logger'
+import MigrateCaptainDuckDuck from '../../utils/MigrateCaptainDuckDuck'
+import Utils from '../../utils/Utils'
+import Authenticator from '../Authenticator'
+import ServiceManager from '../ServiceManager'
+import BackupManager from './BackupManager'
+import CertbotManager from './CertbotManager'
+import DomainResolveChecker from './DomainResolveChecker'
+import LoadBalancerManager from './LoadBalancerManager'
+import SelfHostedDockerRegistry from './SelfHostedDockerRegistry'
 import request = require('request')
 import fs = require('fs-extra')
-import CaptainConstants = require('../../utils/CaptainConstants')
-import Logger = require('../../utils/Logger')
-import LoadBalancerManager = require('./LoadBalancerManager')
-import EnvVars = require('../../utils/EnvVars')
-import CertbotManager = require('./CertbotManager')
-import SelfHostedDockerRegistry = require('./SelfHostedDockerRegistry')
-import ApiStatusCodes = require('../../api/ApiStatusCodes')
-import DataStoreProvider = require('../../datastore/DataStoreProvider')
-import DataStore = require('../../datastore/DataStore')
-import DockerApi from '../../docker/DockerApi'
-import { IRegistryTypes, IRegistryInfo } from '../../models/IRegistryInfo'
-import MigrateCaptainDuckDuck from '../../utils/MigrateCaptainDuckDuck'
-import Authenticator = require('../Authenticator')
-import BackupManager from './BackupManager'
-import ServiceManager = require('../ServiceManager')
-import Utils from '../../utils/Utils'
-import DomainResolveChecker from './DomainResolveChecker'
 
 const DEBUG_SALT = 'THIS IS NOT A REAL CERTIFICATE'
 
@@ -79,17 +78,16 @@ class CaptainManager {
         const dataStore = this.dataStore
         const dockerApi = this.dockerApi
         const loadBalancerManager = this.loadBalancerManager
-        const certbotManager = this.certbotManager
         let myNodeId: string
 
         self.refreshForceSslState()
-            .then(function() {
+            .then(function () {
                 return dockerApi.getNodeIdByServiceName(
                     CaptainConstants.captainServiceName,
                     0
                 )
             })
-            .then(function(nodeId) {
+            .then(function (nodeId) {
                 myNodeId = nodeId
                 self.myNodeId = myNodeId
                 self.dockerRegistry = new SelfHostedDockerRegistry(
@@ -101,47 +99,47 @@ class CaptainManager {
                 )
                 return dockerApi.isNodeManager(myNodeId)
             })
-            .then(function(isManager) {
+            .then(function (isManager) {
                 if (!isManager) {
                     throw new Error('Captain should only run on a manager node')
                 }
             })
-            .then(function() {
+            .then(function () {
                 Logger.d('Emptying generated and temp folders.')
 
                 return fs.emptyDir(CaptainConstants.captainRootDirectoryTemp)
             })
-            .then(function() {
+            .then(function () {
                 return fs.emptyDir(
                     CaptainConstants.captainRootDirectoryGenerated
                 )
             })
-            .then(function() {
+            .then(function () {
                 Logger.d('Ensuring directories are available on host. Started.')
 
                 return fs.ensureDir(CaptainConstants.letsEncryptEtcPath)
             })
-            .then(function() {
+            .then(function () {
                 return fs.ensureDir(CaptainConstants.letsEncryptLibPath)
             })
-            .then(function() {
+            .then(function () {
                 return fs.ensureDir(CaptainConstants.captainStaticFilesDir)
             })
-            .then(function() {
+            .then(function () {
                 return fs.ensureDir(CaptainConstants.perAppNginxConfigPathBase)
             })
-            .then(function() {
+            .then(function () {
                 return fs.ensureFile(CaptainConstants.baseNginxConfigPath)
             })
-            .then(function() {
+            .then(function () {
                 return fs.ensureDir(CaptainConstants.registryPathOnHost)
             })
-            .then(function() {
+            .then(function () {
                 return dockerApi.ensureOverlayNetwork(
                     CaptainConstants.captainNetworkName
                 )
             })
-            .then(function() {
+            .then(function () {
                 Logger.d(
                     'Ensuring directories are available on host. Finished.'
                 )
@@ -151,7 +149,7 @@ class CaptainManager {
                     CaptainConstants.captainNetworkName
                 )
             })
-            .then(function() {
+            .then(function () {
                 const valueIfNotExist = CaptainConstants.isDebug
                     ? DEBUG_SALT
                     : uuid()
@@ -160,29 +158,27 @@ class CaptainManager {
                     valueIfNotExist
                 )
             })
-            .then(function() {
+            .then(function () {
                 return dockerApi.ensureSecretOnService(
                     CaptainConstants.captainServiceName,
                     CaptainConstants.captainSaltSecretKey
                 )
             })
-            .then(function(secretHadExistedBefore) {
+            .then(function (secretHadExistedBefore) {
                 if (!secretHadExistedBefore) {
-                    return new Promise<void>(function() {
+                    return new Promise<void>(function () {
                         Logger.d(
                             'I am halting here. I expect to get restarted in a few seconds due to a secret (captain salt) being updated.'
                         )
                     })
                 }
             })
-            .then(function() {
-                const secretFileName =
-                    '/run/secrets/' + CaptainConstants.captainSaltSecretKey
+            .then(function () {
+                const secretFileName = `/run/secrets/${CaptainConstants.captainSaltSecretKey}`
 
                 if (!fs.pathExistsSync(secretFileName)) {
                     throw new Error(
-                        'Secret is attached according to Docker. But file cannot be found. ' +
-                            secretFileName
+                        `Secret is attached according to Docker. But file cannot be found. ${secretFileName}`
                     )
                 }
 
@@ -196,34 +192,31 @@ class CaptainManager {
 
                 return true
             })
-            .then(function() {
+            .then(function () {
                 return Authenticator.setMainSalt(self.getCaptainSalt())
             })
-            .then(function() {
+            .then(function () {
                 return dataStore.setEncryptionSalt(self.getCaptainSalt())
             })
-            .then(function() {
-                return loadBalancerManager.init(myNodeId, dataStore)
-            })
-            .then(function() {
+            .then(function () {
                 return new MigrateCaptainDuckDuck(
                     dataStore,
                     Authenticator.getAuthenticator(dataStore.getNameSpace())
                 )
                     .migrateIfNeeded()
-                    .then(function(migrationPerformed) {
+                    .then(function (migrationPerformed) {
                         if (!!migrationPerformed) {
                             return self.resetSelf()
                         }
                     })
             })
-            .then(function() {
-                return certbotManager.init(myNodeId)
+            .then(function () {
+                return loadBalancerManager.init(myNodeId, dataStore)
             })
-            .then(function() {
+            .then(function () {
                 return dataStore.getRegistriesDataStore().getAllRegistries()
             })
-            .then(function(registries) {
+            .then(function (registries) {
                 let localRegistry: IRegistryInfo | undefined = undefined
 
                 for (let idx = 0; idx < registries.length; idx++) {
@@ -242,7 +235,7 @@ class CaptainManager {
 
                 return Promise.resolve(true)
             })
-            .then(function() {
+            .then(function () {
                 return self.backupManager.startRestorationIfNeededPhase2(
                     self.getCaptainSalt(),
                     () => {
@@ -250,7 +243,7 @@ class CaptainManager {
                     }
                 )
             })
-            .then(function() {
+            .then(function () {
                 self.inited = true
 
                 self.performHealthCheck()
@@ -259,10 +252,10 @@ class CaptainManager {
                     '**** Captain is initialized and ready to serve you! ****'
                 )
             })
-            .catch(function(error) {
+            .catch(function (error) {
                 Logger.e(error)
 
-                setTimeout(function() {
+                setTimeout(function () {
                     process.exit(0)
                 }, 5000)
             })
@@ -274,14 +267,13 @@ class CaptainManager {
 
     performHealthCheck() {
         const self = this
-        const captainPublicDomain =
-            CaptainConstants.captainSubDomain +
-            '.' +
-            self.dataStore.getRootDomain()
+        const captainPublicDomain = `${
+            CaptainConstants.captainSubDomain
+        }.${self.dataStore.getRootDomain()}`
 
         function scheduleNextHealthCheck() {
             self.healthCheckUuid = uuid()
-            setTimeout(function() {
+            setTimeout(function () {
                 self.performHealthCheck()
             }, HEALTH_CHECK_INTERVAL)
         }
@@ -295,7 +287,7 @@ class CaptainManager {
         function checkCaptainHealth(callback: ISuccessCallback) {
             let callbackCalled = false
 
-            setTimeout(function() {
+            setTimeout(function () {
                 if (callbackCalled) {
                     return
                 }
@@ -305,7 +297,7 @@ class CaptainManager {
             }, TIMEOUT_HEALTH_CHECK)
 
             if (CaptainConstants.configs.skipVerifyingDomains) {
-                setTimeout(function() {
+                setTimeout(function () {
                     if (callbackCalled) {
                         return
                     }
@@ -315,15 +307,12 @@ class CaptainManager {
                 return
             }
 
-            const url =
-                'http://' +
-                captainPublicDomain +
-                CaptainConstants.healthCheckEndPoint
+            const url = `http://${captainPublicDomain}${CaptainConstants.healthCheckEndPoint}`
 
             request(
                 url,
 
-                function(error, response, body) {
+                function (error, response, body) {
                     if (callbackCalled) {
                         return
                     }
@@ -341,7 +330,7 @@ class CaptainManager {
         function checkNginxHealth(callback: ISuccessCallback) {
             let callbackCalled = false
 
-            setTimeout(function() {
+            setTimeout(function () {
                 if (callbackCalled) {
                     return
                 }
@@ -355,7 +344,7 @@ class CaptainManager {
                     captainPublicDomain,
                     '-healthcheck'
                 )
-                .then(function() {
+                .then(function () {
                     if (callbackCalled) {
                         return
                     }
@@ -363,7 +352,7 @@ class CaptainManager {
 
                     callback(true)
                 })
-                .catch(function() {
+                .catch(function () {
                     if (callbackCalled) {
                         return
                     }
@@ -392,18 +381,14 @@ class CaptainManager {
 
             if (!checksPerformed.captainHealth.value) {
                 Logger.w(
-                    'Captain health check failed: #' +
-                        self.consecutiveHealthCheckFailCount +
-                        ' at ' +
-                        captainPublicDomain
+                    `Captain health check failed: #${self.consecutiveHealthCheckFailCount} at ${captainPublicDomain}`
                 )
                 hasFailedCheck = true
             }
 
             if (!checksPerformed.nginxHealth.value) {
                 Logger.w(
-                    'NGINX health check failed: #' +
-                        self.consecutiveHealthCheckFailCount
+                    `NGINX health check failed: #${self.consecutiveHealthCheckFailCount}`
                 )
                 hasFailedCheck = true
             }
@@ -422,14 +407,14 @@ class CaptainManager {
             }
         }
 
-        checkCaptainHealth(function(success) {
+        checkCaptainHealth(function (success) {
             checksPerformed.captainHealth = {
                 value: success,
             }
             scheduleIfNecessary()
         })
 
-        checkNginxHealth(function(success) {
+        checkNginxHealth(function (success) {
             checksPerformed.nginxHealth = {
                 value: success,
             }
@@ -460,10 +445,10 @@ class CaptainManager {
     ensureAllAppsInited() {
         const self = this
         return Promise.resolve() //
-            .then(function() {
+            .then(function () {
                 return self.dataStore.getAppsDataStore().getAppDefinitions()
             })
-            .then(function(apps) {
+            .then(function (apps) {
                 const promises: (() => Promise<void>)[] = []
                 const serviceManager = ServiceManager.get(
                     self.dataStore.getNameSpace(),
@@ -475,18 +460,17 @@ class CaptainManager {
                     CaptainManager.get().getLoadBalanceManager(),
                     CaptainManager.get().getDomainResolveChecker()
                 )
-                Object.keys(apps).forEach(appName => {
-                    promises.push(function() {
+                Object.keys(apps).forEach((appName) => {
+                    promises.push(function () {
                         return Promise.resolve() //
-                            .then(function() {
+                            .then(function () {
                                 return serviceManager.ensureServiceInitedAndUpdated(
                                     appName
                                 )
                             })
-                            .then(function() {
+                            .then(function () {
                                 Logger.d(
-                                    'Waiting 5 second for the service to settle... ' +
-                                        appName
+                                    `Waiting 5 second for the service to settle... ${appName}`
                                 )
                                 return Utils.getDelayedPromise(5000)
                             })
@@ -495,67 +479,6 @@ class CaptainManager {
 
                 return Utils.runPromises(promises)
             })
-    }
-
-    getCaptainImageTags() {
-        const url =
-            'https://hub.docker.com/v2/repositories/' +
-            CaptainConstants.configs.publishedNameOnDockerHub +
-            '/tags'
-
-        return new Promise<string[]>(function(resolve, reject) {
-            request(
-                url,
-
-                function(error, response, body) {
-                    if (CaptainConstants.isDebug) {
-                        resolve(['v0.0.1'])
-                        return
-                    }
-
-                    if (error) {
-                        reject(error)
-                    } else if (!body || !JSON.parse(body).results) {
-                        reject(
-                            new Error(
-                                'Received empty body or no result for version list on docker hub.'
-                            )
-                        )
-                    } else {
-                        const results = JSON.parse(body).results
-                        const tags = []
-                        for (let idx = 0; idx < results.length; idx++) {
-                            tags.push(results[idx].name)
-                        }
-                        resolve(tags)
-                    }
-                }
-            )
-        })
-    }
-
-    updateCaptain(versionTag: string) {
-        const self = this
-        return Promise.resolve().then(function() {
-            return self.dockerApi.updateService(
-                CaptainConstants.captainServiceName,
-                CaptainConstants.configs.publishedNameOnDockerHub +
-                    ':' +
-                    versionTag,
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                undefined
-            )
-        })
     }
 
     getMyNodeId() {
@@ -583,13 +506,13 @@ class CaptainManager {
         const dockerApi = this.dockerApi
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 return dockerApi.ensureContainerStoppedAndRemoved(
                     CaptainConstants.netDataContainerName,
                     CaptainConstants.captainNetworkName
                 )
             })
-            .then(function() {
+            .then(function () {
                 if (netDataInfo.isEnabled) {
                     const vols = [
                         {
@@ -696,7 +619,7 @@ class CaptainManager {
                 // Just removing the old container. No need to create a new one.
                 return true
             })
-            .then(function() {
+            .then(function () {
                 return self.dataStore.setNetDataInfo(netDataInfo)
             })
     }
@@ -705,10 +628,10 @@ class CaptainManager {
         const dockerApi = this.dockerApi
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 return dockerApi.getNodesInfo()
             })
-            .then(function(data) {
+            .then(function (data) {
                 if (!data || !data.length) {
                     throw ApiStatusCodes.createError(
                         ApiStatusCodes.STATUS_ERROR_GENERIC,
@@ -724,16 +647,6 @@ class CaptainManager {
         return this.loadBalancerManager
     }
 
-    reloadLoadBalancer(datastore: DataStore) {
-        const self = this
-        return self.loadBalancerManager
-            .rePopulateNginxConfigFile(datastore)
-            .then(function() {
-                Logger.d('sendReloadSignal...')
-                return self.loadBalancerManager.sendReloadSignal()
-            })
-    }
-
     getDockerRegistry() {
         return this.dockerRegistry
     }
@@ -741,34 +654,37 @@ class CaptainManager {
     enableSsl(emailAddress: string) {
         const self = this
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 return self.certbotManager.ensureRegistered(emailAddress)
             })
-            .then(function() {
+            .then(function () {
                 return self.certbotManager.enableSsl(
-                    CaptainConstants.captainSubDomain +
-                        '.' +
-                        self.dataStore.getRootDomain()
+                    `${
+                        CaptainConstants.captainSubDomain
+                    }.${self.dataStore.getRootDomain()}`
                 )
             })
-            .then(function() {
+            .then(function () {
                 return self.dataStore.setUserEmailAddress(emailAddress)
             })
-            .then(function() {
+            .then(function () {
                 return self.dataStore.setHasRootSsl(true)
             })
-            .then(function() {
-                return self.reloadLoadBalancer(self.dataStore)
+            .then(function () {
+                Logger.d('Updating Load Balancer - CaptainManager.enableSsl')
+                return self.loadBalancerManager.rePopulateNginxConfigFile(
+                    self.dataStore
+                )
             })
     }
 
     forceSsl(isEnabled: boolean) {
         const self = this
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 return self.dataStore.getHasRootSsl()
             })
-            .then(function(hasRootSsl) {
+            .then(function (hasRootSsl) {
                 if (!hasRootSsl && isEnabled) {
                     throw ApiStatusCodes.createError(
                         ApiStatusCodes.STATUS_ERROR_GENERIC,
@@ -778,7 +694,7 @@ class CaptainManager {
 
                 return self.dataStore.setForceSsl(isEnabled)
             })
-            .then(function() {
+            .then(function () {
                 return self.refreshForceSslState()
             })
     }
@@ -786,10 +702,10 @@ class CaptainManager {
     refreshForceSslState() {
         const self = this
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 return self.dataStore.getForceSsl()
             })
-            .then(function(hasForceSsl) {
+            .then(function (hasForceSsl) {
                 self.hasForceSsl = hasForceSsl
             })
     }
@@ -800,7 +716,7 @@ class CaptainManager {
 
     getNginxConfig() {
         const self = this
-        return Promise.resolve().then(function() {
+        return Promise.resolve().then(function () {
             return self.dataStore.getNginxConfig()
         })
     }
@@ -808,10 +724,10 @@ class CaptainManager {
     setNginxConfig(baseConfig: string, captainConfig: string) {
         const self = this
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 return self.dataStore.setNginxConfig(baseConfig, captainConfig)
             })
-            .then(function() {
+            .then(function () {
                 self.resetSelf()
             })
     }
@@ -821,19 +737,16 @@ class CaptainManager {
         // Some DNS servers do not allow wild cards. Therefore this line may fail.
         // We still allow users to specify the domains in their DNS settings individually
         // SubDomains that need to be added are "captain." "registry." "app-name."
-        const url =
-            uuid() +
-            '.' +
-            requestedCustomDomain +
-            ':' +
+        const url = `${uuid()}.${requestedCustomDomain}:${
             CaptainConstants.nginxPortNumber
+        }`
 
         return self.domainResolveChecker
             .verifyDomainResolvesToDefaultServerOnHost(url)
-            .then(function() {
+            .then(function () {
                 return self.dataStore.getHasRootSsl()
             })
-            .then(function(hasRootSsl) {
+            .then(function (hasRootSsl) {
                 if (
                     !force &&
                     hasRootSsl &&
@@ -848,21 +761,50 @@ class CaptainManager {
                 if (force) {
                     return self
                         .forceSsl(false)
-                        .then(function() {
+                        .then(function () {
                             return self.dataStore.setHasRootSsl(false)
                         })
-                        .then(function() {
+                        .then(function () {
                             return self.dataStore
                                 .getAppsDataStore()
                                 .ensureAllAppsSubDomainSslDisabled()
                         })
                 }
             })
-            .then(function() {
+            .then(function () {
+                return self.dataStore
+                    .getRegistriesDataStore()
+                    .getAllRegistries()
+            })
+            .then(function (registries) {
+                let localRegistry: IRegistryInfo | undefined = undefined
+
+                for (let idx = 0; idx < registries.length; idx++) {
+                    const element = registries[idx]
+                    if (element.registryType === IRegistryTypes.LOCAL_REG) {
+                        localRegistry = element
+                    }
+                }
+
+                if (!!localRegistry) {
+                    throw ApiStatusCodes.createError(
+                        ApiStatusCodes.ILLEGAL_OPERATION,
+                        'Delete your self-hosted Docker registry before changing the domain.'
+                    )
+                }
+
+                return Promise.resolve(true)
+            })
+            .then(function () {
                 return self.dataStore.setCustomDomain(requestedCustomDomain)
             })
-            .then(function() {
-                return self.reloadLoadBalancer(self.dataStore)
+            .then(function () {
+                Logger.d(
+                    'Updating Load Balancer - CaptainManager.changeCaptainRootDomain'
+                )
+                return self.loadBalancerManager.rePopulateNginxConfigFile(
+                    self.dataStore
+                )
             })
     }
 
@@ -870,10 +812,11 @@ class CaptainManager {
         const self = this
         Logger.d('Captain is resetting itself!')
         self.waitUntilRestarted = true
-        return new Promise<void>(function(resolve, reject) {
-            setTimeout(function() {
-                const promiseToIgnore = self.dockerApi.updateService(
+        return new Promise<void>(function (resolve, reject) {
+            setTimeout(function () {
+                return self.dockerApi.updateService(
                     CaptainConstants.captainServiceName,
+                    undefined,
                     undefined,
                     undefined,
                     undefined,
@@ -902,4 +845,4 @@ class CaptainManager {
     }
 }
 
-export = CaptainManager
+export default CaptainManager
