@@ -1,10 +1,9 @@
-import CaptainConstants = require('../../utils/CaptainConstants')
-import Logger = require('../../utils/Logger')
-import fs = require('fs-extra')
-import uuid = require('uuid/v4')
-import ApiStatusCodes = require('../../api/ApiStatusCodes')
+import ApiStatusCodes from '../../api/ApiStatusCodes'
 import DockerApi from '../../docker/DockerApi'
+import CaptainConstants from '../../utils/CaptainConstants'
+import Logger from '../../utils/Logger'
 import Utils from '../../utils/Utils'
+import fs = require('fs-extra')
 
 const WEBROOT_PATH_IN_CERTBOT = '/captain-webroot'
 const WEBROOT_PATH_IN_CAPTAIN =
@@ -37,7 +36,7 @@ class CertbotManager {
 
         self.domainValidOrThrow(domainName)
 
-        return '/live/' + domainName + '/fullchain.pem'
+        return `/live/${domainName}/fullchain.pem`
     }
 
     getKeyRelativePathForDomain(domainName: string) {
@@ -45,27 +44,26 @@ class CertbotManager {
 
         self.domainValidOrThrow(domainName)
 
-        return '/live/' + domainName + '/privkey.pem'
+        return `/live/${domainName}/privkey.pem`
     }
 
     enableSsl(domainName: string) {
-        const dockerApi = this.dockerApi
         const self = this
 
-        Logger.d('Enabling SSL for ' + domainName)
+        Logger.d(`Enabling SSL for ${domainName}`)
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 self.domainValidOrThrow(domainName)
                 return self.ensureDomainHasDirectory(domainName)
             })
-            .then(function() {
+            .then(function () {
                 const cmd = [
                     'certbot',
                     'certonly',
                     '--webroot',
                     '-w',
-                    WEBROOT_PATH_IN_CERTBOT + '/' + domainName,
+                    `${WEBROOT_PATH_IN_CERTBOT}/${domainName}`,
                     '-d',
                     domainName,
                 ]
@@ -74,7 +72,7 @@ class CertbotManager {
                     cmd.push('--staging')
                 }
 
-                return self.runCommand(cmd).then(function(output) {
+                return self.runCommand(cmd).then(function (output) {
                     Logger.d(output)
 
                     if (
@@ -95,21 +93,17 @@ class CertbotManager {
 
                     throw ApiStatusCodes.createError(
                         ApiStatusCodes.VERIFICATION_FAILED,
-                        'Unexpected output when enabling SSL for' +
-                            domainName +
-                            ' with ACME Certbot \n' +
-                            output
+                        `Unexpected output when enabling SSL for${domainName} with ACME Certbot \n ${output}`
                     )
                 })
             })
     }
 
     ensureRegistered(emailAddress: string) {
-        const dockerApi = this.dockerApi
         const self = this
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 // Creds used to be saved at
                 // /etc/letencrypt/accounts/acme-v01.api.letsencrypt.org/directory/9fc95dbca2f0b877
                 // After moving to 0.29.1, Certbot started using v2 API. and this path is no longer valid.
@@ -129,7 +123,7 @@ class CertbotManager {
 
                 return self.runCommand(cmd)
             })
-            .then(function(registerOutput) {
+            .then(function (registerOutput) {
                 if (
                     registerOutput.indexOf(
                         'Your account credentials have been saved in your Certbot'
@@ -145,8 +139,7 @@ class CertbotManager {
                 }
 
                 throw new Error(
-                    'Unexpected output when registering with ACME Certbot \n' +
-                        registerOutput
+                    `Unexpected output when registering with ACME Certbot \n ${registerOutput}`
                 )
             })
     }
@@ -182,13 +175,13 @@ class CertbotManager {
     ensureAllCurrentlyRegisteredDomainsHaveDirs() {
         const self = this
         return Promise.resolve() //
-            .then(function() {
+            .then(function () {
                 return self
                     .runCommand(['certbot', 'certificates'])
-                    .then(function(output) {
+                    .then(function (output) {
                         const lines = output.split('\n')
                         const domains: string[] = []
-                        lines.forEach(l => {
+                        lines.forEach((l) => {
                             if (l.indexOf('Certificate Name:') >= 0) {
                                 domains.push(
                                     l.replace('Certificate Name:', '').trim()
@@ -199,10 +192,10 @@ class CertbotManager {
                         return domains
                     })
             })
-            .then(function(allDomains) {
+            .then(function (allDomains) {
                 const p = Promise.resolve()
-                allDomains.forEach(d => {
-                    p.then(function() {
+                allDomains.forEach((d) => {
+                    p.then(function () {
                         return self.ensureDomainHasDirectory(d)
                     })
                 })
@@ -230,7 +223,7 @@ class CertbotManager {
         const dockerApi = this.dockerApi
         const self = this
 
-        return Promise.resolve().then(function() {
+        return Promise.resolve().then(function () {
             self.lock()
 
             const nonInterActiveCommand = [...cmd, '--non-interactive']
@@ -239,11 +232,12 @@ class CertbotManager {
                     CaptainConstants.certbotServiceName,
                     nonInterActiveCommand
                 )
-                .then(function(data) {
+                .then(function (data) {
                     self.unlock()
+                    Logger.dev(data)
                     return data
                 })
-                .catch(function(error) {
+                .catch(function (error) {
                     self.unlock()
                     throw error
                 })
@@ -252,8 +246,8 @@ class CertbotManager {
 
     ensureDomainHasDirectory(domainName: string) {
         return Promise.resolve() //
-            .then(function() {
-                return fs.ensureDir(WEBROOT_PATH_IN_CAPTAIN + '/' + domainName)
+            .then(function () {
+                return fs.ensureDir(`${WEBROOT_PATH_IN_CAPTAIN}/${domainName}`)
             })
     }
 
@@ -271,14 +265,6 @@ class CertbotManager {
             it can be run as frequently as you want - since it will usually take no action.
          */
 
-        // before doing renewal, let's schedule the next one in 20.3 hours!
-        // this random schedule helps to avoid retrying at the same time of
-        // the day in case if that's our super high traffic time
-
-        setTimeout(function() {
-            self.renewAllCerts()
-        }, 1000 * 3600 * 20.3)
-
         const cmd = ['certbot', 'renew']
 
         if (shouldUseStaging) {
@@ -286,16 +272,16 @@ class CertbotManager {
         }
 
         return Promise.resolve() //
-            .then(function() {
+            .then(function () {
                 return self.ensureAllCurrentlyRegisteredDomainsHaveDirs()
             })
-            .then(function() {
+            .then(function () {
                 return self.runCommand(cmd)
             })
-            .then(function(output) {
+            .then(function (output) {
                 // Ignore output :)
             })
-            .catch(function(err) {
+            .catch(function (err) {
                 Logger.e(err)
             })
     }
@@ -317,28 +303,28 @@ class CertbotManager {
                     undefined,
                     undefined
                 )
-                .then(function() {
+                .then(function () {
                     Logger.d('Waiting for Certbot...')
                     return Utils.getDelayedPromise(12000)
                 })
         }
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 return fs.ensureDir(CaptainConstants.letsEncryptEtcPath)
             })
-            .then(function() {
+            .then(function () {
                 return fs.ensureDir(CaptainConstants.letsEncryptLibPath)
             })
-            .then(function() {
+            .then(function () {
                 return fs.ensureDir(WEBROOT_PATH_IN_CAPTAIN)
             })
-            .then(function() {
+            .then(function () {
                 return dockerApi.isServiceRunningByName(
                     CaptainConstants.certbotServiceName
                 )
             })
-            .then(function(isRunning) {
+            .then(function (isRunning) {
                 if (isRunning) {
                     Logger.d('Captain Certbot is already running.. ')
 
@@ -352,12 +338,12 @@ class CertbotManager {
                     )
 
                     return createCertbotServiceOnNode(myNodeId) //
-                        .then(function() {
+                        .then(function () {
                             return myNodeId
                         })
                 }
             })
-            .then(function(nodeId) {
+            .then(function (nodeId) {
                 if (nodeId !== myNodeId) {
                     Logger.d(
                         'Captain Certbot is running on a different node. Removing...'
@@ -367,13 +353,13 @@ class CertbotManager {
                         .removeServiceByName(
                             CaptainConstants.certbotServiceName
                         )
-                        .then(function() {
+                        .then(function () {
                             Logger.d('Waiting for Certbot to be removed...')
                             return Utils.getDelayedPromise(10000)
                         })
-                        .then(function() {
+                        .then(function () {
                             return createCertbotServiceOnNode(myNodeId).then(
-                                function() {
+                                function () {
                                     return true
                                 }
                             )
@@ -382,7 +368,7 @@ class CertbotManager {
                     return true
                 }
             })
-            .then(function() {
+            .then(function () {
                 Logger.d('Updating Certbot service...')
 
                 return dockerApi.updateService(
@@ -413,19 +399,14 @@ class CertbotManager {
                     undefined,
                     undefined,
                     undefined,
+                    undefined,
                     undefined
                 )
             })
-            .then(function() {
+            .then(function () {
                 return self.ensureAllCurrentlyRegisteredDomainsHaveDirs()
-            })
-            .then(function() {
-                // schedule the first attempt to renew certs in 1 minute
-                setTimeout(function() {
-                    self.renewAllCerts()
-                }, 1000 * 60)
             })
     }
 }
 
-export = CertbotManager
+export default CertbotManager
