@@ -1,17 +1,19 @@
 import ejs = require('ejs')
-import CaptainConstants = require('../../utils/CaptainConstants')
-import Logger = require('../../utils/Logger')
-import fs = require('fs-extra')
-import uuid = require('uuid/v4')
-import request = require('request')
-import ApiStatusCodes = require('../../api/ApiStatusCodes')
-import DockerApi from '../../docker/DockerApi'
-import DataStore = require('../../datastore/DataStore')
-import CertbotManager = require('./CertbotManager')
-import { AnyError } from '../../models/OtherTypes'
-import LoadBalancerInfo from '../../models/LoadBalancerInfo'
+import * as chileProcess from 'child_process'
 import * as path from 'path'
-import Utils from '../../utils/Utils'
+import * as util from 'util'
+import { v4 as uuid } from 'uuid'
+import ApiStatusCodes from '../../api/ApiStatusCodes'
+import DataStore from '../../datastore/DataStore'
+import DockerApi from '../../docker/DockerApi'
+import LoadBalancerInfo from '../../models/LoadBalancerInfo'
+import { AnyError } from '../../models/OtherTypes'
+import CaptainConstants from '../../utils/CaptainConstants'
+import Logger from '../../utils/Logger'
+import CertbotManager from './CertbotManager'
+import fs = require('fs-extra')
+import request = require('request')
+const exec = util.promisify(chileProcess.exec)
 
 const defaultPageTemplate = fs
     .readFileSync(__dirname + '/../../../template/default-page.ejs')
@@ -29,6 +31,16 @@ const HOST_PATH_OF_FAKE_CERTS =
 if (!fs.existsSync(CAPROVER_CONTAINER_PATH_OF_FAKE_CERTS))
     throw new Error('CAPROVER_CONTAINER_PATH_OF_FAKE_CERTS  is empty')
 if (!defaultPageTemplate) throw new Error('defaultPageTemplate  is empty')
+
+const DH_PARAMS_FILE_PATH_ON_HOST = path.join(
+    CaptainConstants.nginxSharedPathOnHost,
+    CaptainConstants.nginxDhParamFileName
+)
+
+const DH_PARAMS_FILE_PATH_ON_NGINX = path.join(
+    CaptainConstants.nginxSharedPathOnNginx,
+    CaptainConstants.nginxDhParamFileName
+)
 
 class LoadBalancerManager {
     private reloadInProcess: boolean
@@ -55,16 +67,22 @@ class LoadBalancerManager {
      * @param dataStoreToQueue
      * @returns {Promise.<>}
      */
-    rePopulateNginxConfigFile(dataStoreToQueue: DataStore) {
+    rePopulateNginxConfigFile(dataStoreToQueue: DataStore, noReload?: boolean) {
         const self = this
 
-        return new Promise<void>(function(res, rej) {
+        return new Promise<void>(function (res, rej) {
             self.requestedReloadPromises.push({
                 dataStore: dataStoreToQueue,
                 resolve: res,
                 reject: rej,
             })
             self.consumeQueueIfAnyInNginxReloadQueue()
+        }).then(function () {
+            if (!!noReload) return
+            Logger.d('sendReloadSignal...')
+            return self.dockerApi.sendSingleContainerKillHUP(
+                CaptainConstants.nginxServiceName
+            )
         })
     }
 
@@ -89,10 +107,9 @@ class LoadBalancerManager {
         const dataStore = q.dataStore
 
         // This will resolve to something like: /captain/nginx/conf.d/captain
-        const configFilePathBase =
-            CaptainConstants.perAppNginxConfigPathBase +
-            '/' +
-            dataStore.getNameSpace()
+        const configFilePathBase = `${
+            CaptainConstants.perAppNginxConfigPathBase
+        }/${dataStore.getNameSpace()}`
 
         const FUTURE = configFilePathBase + '.fut'
         const BACKUP = configFilePathBase + '.bak'
@@ -101,13 +118,13 @@ class LoadBalancerManager {
         let nginxConfigContent = ''
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 return fs.remove(FUTURE)
             })
-            .then(function() {
+            .then(function () {
                 return self.getServerList(dataStore)
             })
-            .then(function(servers) {
+            .then(function (servers) {
                 const promises: Promise<void>[] = []
 
                 if (servers && !!servers.length) {
@@ -118,22 +135,20 @@ class LoadBalancerManager {
                             s.keyPath = self.getSslKeyPath(s.publicDomain)
                         }
 
-                        s.staticWebRoot =
+                        s.staticWebRoot = `${
                             CaptainConstants.nginxStaticRootDir +
-                            CaptainConstants.nginxDomainSpecificHtmlDir +
-                            '/' +
-                            s.publicDomain
+                            CaptainConstants.nginxDomainSpecificHtmlDir
+                        }/${s.publicDomain}`
 
                         s.customErrorPagesDirectory =
                             CaptainConstants.nginxStaticRootDir +
                             CaptainConstants.nginxDefaultHtmlDir
 
-                        const pathOfAuthInHost =
-                            configFilePathBase + '-' + s.publicDomain + '.auth'
+                        const pathOfAuthInHost = `${configFilePathBase}-${s.publicDomain}.auth`
 
                         promises.push(
                             Promise.resolve()
-                                .then(function() {
+                                .then(function () {
                                     if (s.httpBasicAuth) {
                                         s.httpBasicAuthPath = path.join(
                                             CONTAINER_PATH_OF_CONFIG,
@@ -145,12 +160,12 @@ class LoadBalancerManager {
                                         )
                                     }
                                 })
-                                .then(function() {
+                                .then(function () {
                                     return ejs.render(s.nginxConfigTemplate, {
                                         s: s,
                                     })
                                 })
-                                .then(function(rendered) {
+                                .then(function (rendered) {
                                     nginxConfigContent += rendered
                                 })
                         )
@@ -159,31 +174,34 @@ class LoadBalancerManager {
 
                 return Promise.all(promises)
             })
-            .then(function() {
+            .then(function () {
                 return fs.outputFile(FUTURE, nginxConfigContent)
             })
-            .then(function() {
+            .then(function () {
                 return fs.remove(BACKUP)
             })
-            .then(function() {
+            .then(function () {
                 return fs.ensureFile(CONFIG)
             })
-            .then(function() {
+            .then(function () {
                 return fs.renameSync(CONFIG, BACKUP) // sync method. It's really fast.
             })
-            .then(function() {
+            .then(function () {
                 return fs.renameSync(FUTURE, CONFIG) // sync method. It's really fast.
             })
-            .then(function() {
+            .then(function () {
+                return self.ensureBaseNginxConf()
+            })
+            .then(function () {
                 return self.createRootConfFile(dataStore)
             })
-            .then(function() {
+            .then(function () {
                 Logger.d('SUCCESS: UNLocking NGINX configuration reloading...')
                 self.reloadInProcess = false
                 q.resolve()
                 self.consumeQueueIfAnyInNginxReloadQueue()
             })
-            .catch(function(error: AnyError) {
+            .catch(function (error: AnyError) {
                 Logger.e(error)
                 Logger.d('Error: UNLocking NGINX configuration reloading...')
                 self.reloadInProcess = false
@@ -199,21 +217,21 @@ class LoadBalancerManager {
         let rootDomain: string
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 return dataStore.getHasRootSsl()
             })
-            .then(function(val: boolean) {
+            .then(function (val: boolean) {
                 hasRootSsl = val
 
                 return dataStore.getRootDomain()
             })
-            .then(function(val) {
+            .then(function (val) {
                 rootDomain = val
             })
-            .then(function() {
+            .then(function () {
                 return dataStore.getDefaultAppNginxConfig()
             })
-            .then(function(defaultAppNginxConfig) {
+            .then(function (defaultAppNginxConfig) {
                 return self.getAppsServerConfig(
                     dataStore,
                     defaultAppNginxConfig,
@@ -229,21 +247,17 @@ class LoadBalancerManager {
         hasRootSsl: boolean,
         rootDomain: string
     ) {
-        const self = this
-
         const servers: IServerBlockDetails[] = []
 
         return dataStore
             .getAppsDataStore()
             .getAppDefinitions()
-            .then(function(apps) {
-                Object.keys(apps).forEach(function(appName) {
+            .then(function (apps) {
+                Object.keys(apps).forEach(function (appName) {
                     const webApp = apps[appName]
                     const httpBasicAuth =
                         webApp.httpAuth && webApp.httpAuth.passwordHashed //
-                            ? webApp.httpAuth.user +
-                              ':' +
-                              webApp.httpAuth.passwordHashed
+                            ? `${webApp.httpAuth.user}:${webApp.httpAuth.passwordHashed}`
                             : ''
 
                     if (webApp.notExposeAsWebApp) {
@@ -261,8 +275,7 @@ class LoadBalancerManager {
                     const serverWithSubDomain = {} as IServerBlockDetails
                     serverWithSubDomain.hasSsl =
                         hasRootSsl && webApp.hasDefaultSubDomainSsl
-                    serverWithSubDomain.publicDomain =
-                        appName + '.' + rootDomain
+                    serverWithSubDomain.publicDomain = `${appName}.${rootDomain}`
                     serverWithSubDomain.localDomain = localDomain
                     serverWithSubDomain.forceSsl = forceSsl
                     serverWithSubDomain.websocketSupport = websocketSupport
@@ -302,12 +315,6 @@ class LoadBalancerManager {
             })
     }
 
-    sendReloadSignal() {
-        return this.dockerApi.sendSingleContainerKillHUP(
-            CaptainConstants.nginxServiceName
-        )
-    }
-
     getCaptainPublicRandomKey() {
         return this.captainPublicRandomKey
     }
@@ -329,13 +336,12 @@ class LoadBalancerManager {
     }
 
     getInfo() {
-        return new Promise<LoadBalancerInfo>(function(resolve, reject) {
-            const url =
-                'http://' + CaptainConstants.nginxServiceName + '/nginx_status'
+        return new Promise<LoadBalancerInfo>(function (resolve, reject) {
+            const url = `http://${CaptainConstants.nginxServiceName}/nginx_status`
 
-            request(url, function(error, response, body) {
+            request(url, function (error, response, body) {
                 if (error || !body) {
-                    Logger.e('Error        ' + error)
+                    Logger.e(`Error        ${error}`)
                     reject(
                         ApiStatusCodes.createError(
                             ApiStatusCodes.STATUS_ERROR_GENERIC,
@@ -378,10 +384,12 @@ class LoadBalancerManager {
     createRootConfFile(dataStore: DataStore) {
         const self = this
 
-        const captainDomain =
-            CaptainConstants.captainSubDomain + '.' + dataStore.getRootDomain()
-        const registryDomain =
-            CaptainConstants.registrySubDomain + '.' + dataStore.getRootDomain()
+        const captainDomain = `${
+            CaptainConstants.captainSubDomain
+        }.${dataStore.getRootDomain()}`
+        const registryDomain = `${
+            CaptainConstants.registrySubDomain
+        }.${dataStore.getRootDomain()}`
 
         let hasRootSsl = false
 
@@ -392,21 +400,21 @@ class LoadBalancerManager {
         let rootNginxTemplate: string | undefined = undefined
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 return dataStore.getNginxConfig()
             })
-            .then(function(nginxConfig) {
+            .then(function (nginxConfig) {
                 rootNginxTemplate =
                     nginxConfig.captainConfig.customValue ||
                     nginxConfig.captainConfig.byDefault
 
                 return dataStore.getHasRootSsl()
             })
-            .then(function(hasSsl) {
+            .then(function (hasSsl) {
                 hasRootSsl = hasSsl
                 return dataStore.getHasRegistrySsl()
             })
-            .then(function(hasRegistrySsl) {
+            .then(function (hasRegistrySsl) {
                 return ejs.render(rootNginxTemplate!, {
                     fake: {
                         crtPath: path.join(
@@ -429,38 +437,36 @@ class LoadBalancerManager {
                         defaultHtmlDir:
                             CaptainConstants.nginxStaticRootDir +
                             CaptainConstants.nginxDefaultHtmlDir,
-                        staticWebRoot:
+                        staticWebRoot: `${
                             CaptainConstants.nginxStaticRootDir +
-                            CaptainConstants.nginxDomainSpecificHtmlDir +
-                            '/' +
-                            captainDomain,
+                            CaptainConstants.nginxDomainSpecificHtmlDir
+                        }/${captainDomain}`,
                     },
                     registry: {
                         crtPath: self.getSslCertPath(registryDomain),
                         keyPath: self.getSslKeyPath(registryDomain),
                         hasRootSsl: hasRegistrySsl,
                         domain: registryDomain,
-                        staticWebRoot:
+                        staticWebRoot: `${
                             CaptainConstants.nginxStaticRootDir +
-                            CaptainConstants.nginxDomainSpecificHtmlDir +
-                            '/' +
-                            registryDomain,
+                            CaptainConstants.nginxDomainSpecificHtmlDir
+                        }/${registryDomain}`,
                     },
                 })
             })
-            .then(function(rootNginxConfContent) {
+            .then(function (rootNginxConfContent) {
                 return fs.outputFile(FUTURE, rootNginxConfContent)
             })
-            .then(function() {
+            .then(function () {
                 return fs.remove(BACKUP)
             })
-            .then(function() {
+            .then(function () {
                 return fs.ensureFile(CONFIG)
             })
-            .then(function() {
+            .then(function () {
                 return fs.renameSync(CONFIG, BACKUP) // sync method. It's really fast.
             })
-            .then(function() {
+            .then(function () {
                 return fs.renameSync(FUTURE, CONFIG) // sync method. It's really fast.
             })
     }
@@ -468,22 +474,77 @@ class LoadBalancerManager {
     ensureBaseNginxConf() {
         const self = this
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 return self.dataStore.getNginxConfig()
             })
-            .then(function(captainConfig) {
+            .then(function (captainConfig) {
                 const baseConfigTemplate =
                     captainConfig.baseConfig.customValue ||
                     captainConfig.baseConfig.byDefault
 
-                return ejs.render(baseConfigTemplate, {})
+                return ejs.render(baseConfigTemplate, {
+                    base: {
+                        dhparamsFilePath:
+                            fs.existsSync(DH_PARAMS_FILE_PATH_ON_HOST) &&
+                            fs
+                                .readFileSync(DH_PARAMS_FILE_PATH_ON_HOST)
+                                .toString().length > 10 // making sure it's not an buggy file
+                                ? DH_PARAMS_FILE_PATH_ON_NGINX
+                                : '',
+                    },
+                })
             })
-            .then(function(baseNginxConfFileContent) {
+            .then(function (baseNginxConfFileContent) {
                 return fs.outputFile(
                     CaptainConstants.baseNginxConfigPath,
                     baseNginxConfFileContent
                 )
             })
+    }
+
+    ensureDhParamFileExists(dataStore: DataStore) {
+        const self = this
+        return fs
+            .pathExists(DH_PARAMS_FILE_PATH_ON_HOST) //
+            .then(function (dhParamExists) {
+                if (!dhParamExists) {
+                    return false
+                }
+
+                const dhFileContent = fs
+                    .readFileSync(DH_PARAMS_FILE_PATH_ON_HOST)
+                    .toString()
+
+                const contentValid =
+                    dhFileContent.indexOf('END DH PARAMETERS') > 0
+
+                if (contentValid) {
+                    return true
+                }
+
+                Logger.d(
+                    `Invalid dh param content - size of: ${dhFileContent.length}`
+                )
+                fs.removeSync(DH_PARAMS_FILE_PATH_ON_HOST)
+
+                return false
+            })
+            .then(function (dhParamExists) {
+                if (dhParamExists) {
+                    return
+                }
+
+                Logger.d(
+                    'Creating dhparams for the first time - high CPU load is expected.'
+                )
+                return exec(
+                    `openssl dhparam -out ${DH_PARAMS_FILE_PATH_ON_HOST} 2048`
+                ).then(function () {
+                    Logger.d('Updating Load Balancer - ensureDhParamFileExists')
+                    return self.rePopulateNginxConfigFile(dataStore)
+                })
+            })
+            .catch((err) => Logger.e(err))
     }
 
     init(myNodeId: string, dataStore: DataStore) {
@@ -522,15 +583,15 @@ class LoadBalancerManager {
                         },
                     }
                 )
-                .then(function() {
+                .then(function () {
                     const waitTimeInMillis = 5000
                     Logger.d(
-                        'Waiting for ' +
-                            waitTimeInMillis / 1000 +
-                            ' seconds for nginx to start up'
+                        `Waiting for ${
+                            waitTimeInMillis / 1000
+                        } seconds for nginx to start up`
                     )
-                    return new Promise<boolean>(function(resolve, reject) {
-                        setTimeout(function() {
+                    return new Promise<boolean>(function (resolve, reject) {
+                        setTimeout(function () {
                             resolve(true)
                         }, waitTimeInMillis)
                     })
@@ -544,7 +605,7 @@ class LoadBalancerManager {
                     CaptainConstants.captainConfirmationPath,
                 self.getCaptainPublicRandomKey()
             )
-            .then(function() {
+            .then(function () {
                 return ejs.render(defaultPageTemplate, {
                     message_title: 'Nothing here yet :/',
                     message_body: '',
@@ -552,7 +613,7 @@ class LoadBalancerManager {
                     message_link_title: 'Read Docs',
                 })
             })
-            .then(function(staticPageContent) {
+            .then(function (staticPageContent) {
                 return fs.outputFile(
                     CaptainConstants.captainStaticFilesDir +
                         CaptainConstants.nginxDefaultHtmlDir +
@@ -560,7 +621,7 @@ class LoadBalancerManager {
                     staticPageContent
                 )
             })
-            .then(function() {
+            .then(function () {
                 return ejs.render(defaultPageTemplate, {
                     message_title: 'An Error Occurred :/',
                     message_body: '',
@@ -568,7 +629,7 @@ class LoadBalancerManager {
                     message_link_title: 'Read Docs',
                 })
             })
-            .then(function(errorGenericPageContent) {
+            .then(function (errorGenericPageContent) {
                 return fs.outputFile(
                     CaptainConstants.captainStaticFilesDir +
                         CaptainConstants.nginxDefaultHtmlDir +
@@ -576,7 +637,7 @@ class LoadBalancerManager {
                     errorGenericPageContent
                 )
             })
-            .then(function() {
+            .then(function () {
                 return ejs.render(defaultPageTemplate, {
                     message_title: 'NGINX 502 Error :/',
                     message_body:
@@ -586,7 +647,7 @@ class LoadBalancerManager {
                     message_link_title: 'Docs - 502 Troubleshooting',
                 })
             })
-            .then(function(error502PageContent) {
+            .then(function (error502PageContent) {
                 return fs.outputFile(
                     CaptainConstants.captainStaticFilesDir +
                         CaptainConstants.nginxDefaultHtmlDir +
@@ -594,7 +655,7 @@ class LoadBalancerManager {
                     error502PageContent
                 )
             })
-            .then(function() {
+            .then(function () {
                 Logger.d('Copying fake certificates...')
 
                 return fs.copy(
@@ -602,26 +663,24 @@ class LoadBalancerManager {
                     HOST_PATH_OF_FAKE_CERTS
                 )
             })
-            .then(function() {
-                Logger.d('Setting up NGINX conf file...')
-
-                return self.ensureBaseNginxConf()
+            .then(function () {
+                Logger.d(
+                    'Updating Load Balancer - Setting up NGINX conf file...'
+                )
+                return self.rePopulateNginxConfigFile(dataStore, true)
             })
-            .then(function() {
-                return self.rePopulateNginxConfigFile(dataStore)
-            })
-            .then(function() {
+            .then(function () {
                 return fs.ensureDir(CaptainConstants.letsEncryptEtcPath)
             })
-            .then(function() {
+            .then(function () {
                 return fs.ensureDir(CaptainConstants.nginxSharedPathOnHost)
             })
-            .then(function() {
+            .then(function () {
                 return dockerApi.isServiceRunningByName(
                     CaptainConstants.nginxServiceName
                 )
             })
-            .then(function(isRunning) {
+            .then(function (isRunning) {
                 if (isRunning) {
                     Logger.d('Captain Nginx is already running.. ')
 
@@ -630,12 +689,12 @@ class LoadBalancerManager {
                         0
                     )
                 } else {
-                    return createNginxServiceOnNode(myNodeId).then(function() {
+                    return createNginxServiceOnNode(myNodeId).then(function () {
                         return myNodeId
                     })
                 }
             })
-            .then(function(nodeId) {
+            .then(function (nodeId) {
                 if (nodeId !== myNodeId) {
                     Logger.d(
                         'Captain Nginx is running on a different node. Removing...'
@@ -643,9 +702,9 @@ class LoadBalancerManager {
 
                     return dockerApi
                         .removeServiceByName(CaptainConstants.nginxServiceName)
-                        .then(function() {
+                        .then(function () {
                             return createNginxServiceOnNode(myNodeId).then(
-                                function() {
+                                function () {
                                     return true
                                 }
                             )
@@ -654,7 +713,7 @@ class LoadBalancerManager {
                     return true
                 }
             })
-            .then(function() {
+            .then(function () {
                 Logger.d('Updating NGINX service...')
 
                 return dockerApi.updateService(
@@ -699,24 +758,58 @@ class LoadBalancerManager {
                     undefined,
                     undefined,
                     undefined,
+                    undefined,
                     undefined
                 )
             })
-            .then(function() {
+            .then(function () {
                 const waitTimeInMillis = 5000
                 Logger.d(
-                    'Waiting for ' +
-                        waitTimeInMillis / 1000 +
-                        ' seconds for nginx reload to take into effect'
+                    `Waiting for ${
+                        waitTimeInMillis / 1000
+                    } seconds for nginx reload to take into effect`
                 )
-                return new Promise<boolean>(function(resolve, reject) {
-                    setTimeout(function() {
+                return new Promise<boolean>(function (resolve, reject) {
+                    setTimeout(function () {
                         Logger.d('NGINX is fully set up and working...')
                         resolve(true)
                     }, waitTimeInMillis)
                 })
             })
+            .then(function () {
+                return self.certbotManager.init(myNodeId)
+            })
+            .then(function () {
+                // schedule the 10sec:
+                // Ensure DH Params exists
+                // First attempt to renew certs in
+                setTimeout(function () {
+                    self.ensureDhParamFileExists(dataStore) //
+                        .then(function () {
+                            return self.renewAllCertsAndReload(dataStore)
+                        })
+                }, 1000 * 10)
+            })
+    }
+
+    renewAllCertsAndReload(dataStore: DataStore) {
+        const self = this
+
+        // before doing renewal, let's schedule the next one in 20.3 hours!
+        // this random schedule helps to avoid retrying at the same time of
+        // the day in case if that's our super high traffic time
+
+        setTimeout(function () {
+            self.renewAllCertsAndReload(dataStore)
+        }, 1000 * 3600 * 20.3)
+
+        return self.certbotManager
+            .renewAllCerts() //
+            .then(function () {
+                Logger.d('Updating Load Balancer - renewAllCerts')
+                return self.rePopulateNginxConfigFile(dataStore)
+            })
     }
 }
 
-export = LoadBalancerManager
+export default LoadBalancerManager
