@@ -1,9 +1,9 @@
 class BaseApi {
     public data: any
 
-    constructor(private status: number, private description: string) {
+    constructor(public status: number, public description: string) {
         this.data = {}
     }
 }
 
-export = BaseApi
+export default BaseApi
