@@ -1,8 +1,13 @@
-import { CaptainEncryptor } from '../utils/Encryptor'
+import { v4 as uuid } from 'uuid'
+import ApiStatusCodes from '../api/ApiStatusCodes'
+import {
+    IRegistryInfo,
+    IRegistryInfoEncrypted,
+    IRegistryType,
+    IRegistryTypes,
+} from '../models/IRegistryInfo'
+import CaptainEncryptor from '../utils/Encryptor'
 import configstore = require('configstore')
-import uuid = require('uuid/v4')
-import ApiStatusCodes = require('../api/ApiStatusCodes')
-import { IRegistryInfo, IRegistryInfoEncrypted, IRegistryTypes, IRegistryType } from '../models/IRegistryInfo'
 
 const DOCKER_REGISTRIES = 'dockerRegistries'
 const DEFAULT_DOCKER_REGISTRY_ID = 'defaultDockerRegId'
@@ -10,7 +15,7 @@ const DEFAULT_DOCKER_REGISTRY_ID = 'defaultDockerRegId'
 class RegistriesDataStore {
     private encryptor: CaptainEncryptor
 
-    constructor(private data: configstore, private namepace: string) {}
+    constructor(private data: configstore, public namepace: string) {}
 
     setEncryptor(encryptor: CaptainEncryptor) {
         this.encryptor = encryptor
@@ -19,7 +24,7 @@ class RegistriesDataStore {
     getDefaultPushRegistryId(): Promise<string | undefined> {
         const self = this
 
-        return Promise.resolve().then(function() {
+        return Promise.resolve().then(function () {
             return self.data.get(DEFAULT_DOCKER_REGISTRY_ID)
         })
     }
@@ -28,10 +33,10 @@ class RegistriesDataStore {
         const self = this
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 return self.getAllRegistries()
             })
-            .then(function(registries) {
+            .then(function (registries) {
                 let found = false
 
                 for (let i = 0; i < registries.length; i++) {
@@ -57,12 +62,12 @@ class RegistriesDataStore {
         const self = this
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 if (!registryId) throw new Error('Empty registry id!')
 
                 return self.getAllRegistries()
             })
-            .then(function(registries) {
+            .then(function (registries) {
                 for (let i = 0; i < registries.length; i++) {
                     const registry = registries[i]
                     if (registry.id === registryId) {
@@ -81,12 +86,12 @@ class RegistriesDataStore {
         const self = this
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 if (!registryId) throw new Error('Empty registry id to delete!')
 
                 return self.getAllRegistries()
             })
-            .then(function(registries) {
+            .then(function (registries) {
                 const newReg = []
                 for (let i = 0; i < registries.length; i++) {
                     const registry = registries[i]
@@ -110,10 +115,10 @@ class RegistriesDataStore {
         const self = this
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 return self.data.get(DOCKER_REGISTRIES) || []
             })
-            .then(function(registries: IRegistryInfoEncrypted[]) {
+            .then(function (registries: IRegistryInfoEncrypted[]) {
                 const unencryptedList: IRegistryInfo[] = []
                 for (let i = 0; i < registries.length; i++) {
                     const element = registries[i]
@@ -142,7 +147,7 @@ class RegistriesDataStore {
         const self = this
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 if (
                     !id ||
                     !registryUser ||
@@ -157,7 +162,7 @@ class RegistriesDataStore {
 
                 return self.getAllRegistries()
             })
-            .then(function(registries) {
+            .then(function (registries) {
                 let found = false
                 for (let idx = 0; idx < registries.length; idx++) {
                     const element = registries[idx]
@@ -199,7 +204,7 @@ class RegistriesDataStore {
         let savedId: string | undefined = undefined
 
         return Promise.resolve()
-            .then(function() {
+            .then(function () {
                 if (
                     !registryUser ||
                     !registryPassword ||
@@ -214,7 +219,7 @@ class RegistriesDataStore {
 
                 return self.getAllRegistries()
             })
-            .then(function(registries) {
+            .then(function (registries) {
                 let id: string = uuid()
                 let isAlreadyTaken = true
 
@@ -241,7 +246,7 @@ class RegistriesDataStore {
 
                 return self.saveAllRegistries(registries)
             })
-            .then(function() {
+            .then(function () {
                 if (!savedId)
                     throw new Error(
                         'Saved registry, but ID is null. This should never happen'
@@ -253,7 +258,7 @@ class RegistriesDataStore {
     private saveAllRegistries(registries: IRegistryInfo[]) {
         const self = this
         return Promise.resolve() //
-            .then(function() {
+            .then(function () {
                 const encryptedList: IRegistryInfoEncrypted[] = []
                 for (let i = 0; i < registries.length; i++) {
                     const element = registries[i]
@@ -273,4 +278,4 @@ class RegistriesDataStore {
     }
 }
 
-export = RegistriesDataStore
+export default RegistriesDataStore
