@@ -1,10 +1,10 @@
-export = {
+export default {
     keys: {
         CAPTAIN_DOCKER_API: 'CAPTAIN_DOCKER_API',
         CAPTAIN_IS_DEBUG: 'CAPTAIN_IS_DEBUG',
         DEFAULT_PASSWORD: 'DEFAULT_PASSWORD',
         IS_CAPTAIN_INSTANCE: 'IS_CAPTAIN_INSTANCE',
-        IS_DEMO_MODE: 'IS_DEMO_MODE',
+        DEMO_MODE_ADMIN_IP: 'DEMO_MODE_ADMIN_IP',
     },
 
     BY_PASS_PROXY_CHECK: process.env.BY_PASS_PROXY_CHECK,
@@ -17,7 +17,7 @@ export = {
 
     IS_CAPTAIN_INSTANCE: process.env.IS_CAPTAIN_INSTANCE,
 
-    IS_DEMO_MODE: !!process.env.IS_DEMO_MODE,
+    DEMO_MODE_ADMIN_IP: process.env.DEMO_MODE_ADMIN_IP,
 
     DEFAULT_PASSWORD: process.env.DEFAULT_PASSWORD,
 }
