@@ -1,4 +1,6 @@
-import { removeSync, remove } from 'fs-extra'
+import { remove } from 'fs-extra'
+import * as yaml from 'yaml'
+import Logger from './Logger'
 
 export default class Utils {
     static removeHttpHttps(input: string) {
@@ -14,8 +16,58 @@ export default class Utils {
         )
     }
 
+    static mergeObjects(object1: any, object2: any) {
+        const newObject = object1 || {}
+        object2 = object2 || {}
+
+        Object.keys(object2).forEach((k) => {
+            if (
+                !newObject[k] ||
+                Array.isArray(newObject[k]) ||
+                Array.isArray(object2[k])
+            ) {
+                newObject[k] = object2[k]
+            } else {
+                if (
+                    typeof object2[k] === 'object' &&
+                    typeof newObject[k] === 'object'
+                ) {
+                    newObject[k] = this.mergeObjects(newObject[k], object2[k])
+                } else {
+                    newObject[k] = object2[k]
+                }
+            }
+        })
+
+        return newObject
+    }
+
+    static convertYamlOrJsonToObject(raw: string | undefined) {
+        raw = !!raw ? `${raw}`.trim() : ''
+        if (!raw.length) {
+            return undefined
+        }
+
+        let returnValue = undefined as any
+        if (raw.startsWith('{') || raw.startsWith('[')) {
+            try {
+                returnValue = JSON.parse(raw)
+            } catch (error) {
+                Logger.e(error)
+            }
+        } else {
+            try {
+                returnValue = yaml.parse(raw)
+            } catch (error) {
+                Logger.e(error)
+            }
+        }
+
+        return returnValue
+    }
+
     static deleteFileQuietly(absFileOrDirPath: string) {
-        return remove(absFileOrDirPath).catch(function(error) {
+        return remove(absFileOrDirPath).catch(function (error) {
             // nom nom
         })
     }
@@ -43,7 +95,7 @@ export default class Utils {
     static filterInPlace<T>(arr: T[], condition: (value: T) => boolean) {
         let newArray = arr.filter(condition)
         arr.splice(0, arr.length)
-        newArray.forEach(value => arr.push(value))
+        newArray.forEach((value) => arr.push(value))
     }
 
     static dropFirstElements(arr: any[], maxLength: number) {
@@ -61,7 +113,7 @@ export default class Utils {
     ): Promise<void> {
         let currCorrected = curr ? curr : 0
         if (promises.length > currCorrected) {
-            return promises[currCorrected]().then(function() {
+            return promises[currCorrected]().then(function () {
                 return Utils.runPromises(promises, currCorrected + 1)
             })
         }
